### 2018-12-21 #1139 @whedon
[PRE REVIEW]: open_iA: A tool for processing and visual analysis of industrial computed tomography datasets
**Submitting author:** @codeling (<a href="http://orcid.org/0000-0003-1271-0838">Bernhard Fröhler</a>)
**Repository:** <a href="https://github.com/3dct/open_iA" target ="_blank">https://github.com/3dct/open_iA</a>
**Version:** 2018.12
