### 2018-12-23 693bceb
Core: Enable selection in standalone scatter plot widget
