+++ core/src/charts/iAScatterPlotWidget.cpp
@@ -90,6 +90,7 @@ iAScatterPlotWidget::iAScatterPlotWidget(QSharedPointer<iASPLOMData> data) :
 	setMouseTracking(true);
 	setFocusPolicy(Qt::StrongFocus);
 	m_scatterplot = new iAScatterPlot(m_scatterPlotHandler.data(), this);
+	m_scatterplot->settings.selectionEnabled = true;
 	data->updateRanges();
 	m_scatterplot->setData(0, 1, data);
 }
