### 2018-12-23 da55f23
Core: Show selection polygon in scatter plot widget
