### 2019-01-02 2d411c3
Reduce Christmas easter egg time, only show in program logo
