### 2019-01-02 f019dcb
VTK 8.2(RC2) compatibility: better conditional compilation with QVTK(OpenGL[Native])Widget(2)
