+++ core/src/QVTKWidgetMouseReleaseWorkaround.cpp
@@ -20,15 +20,9 @@
 * ************************************************************************************/
 #include "QVTKWidgetMouseReleaseWorkaround.h"
 
-#if (VTK_MAJOR_VERSION >= 8 && defined(VTK_OPENGL2_BACKEND) )
 QVTKWidgetMouseReleaseWorkaround::QVTKWidgetMouseReleaseWorkaround(QWidget* parent, Qt::WindowFlags f)
-	: QVTKOpenGLWidget(parent, f)
+	: iAVtkOldWidget(parent, f)
 {}
-#else
-QVTKWidgetMouseReleaseWorkaround::QVTKWidgetMouseReleaseWorkaround(QWidget* parent, Qt::WindowFlags f)
-	: QVTKWidget(parent, f)
-{}
-#endif
 
 void QVTKWidgetMouseReleaseWorkaround::mouseReleaseEvent( QMouseEvent * event )
 {
@@ -36,19 +30,11 @@ void QVTKWidgetMouseReleaseWorkaround::mouseReleaseEvent( QMouseEvent * event )
 		emit rightButtonReleasedSignal();
 	else if ( Qt::LeftButton == event->button() )
 		emit leftButtonReleasedSignal();
-#if (VTK_MAJOR_VERSION >= 8 && defined(VTK_OPENGL2_BACKEND) )
-	QVTKOpenGLWidget::mouseReleaseEvent(event);
-#else
-	QVTKWidget::mouseReleaseEvent(event);
-#endif
+	iAVtkOldWidget::mouseReleaseEvent(event);
 }
 
 void QVTKWidgetMouseReleaseWorkaround::resizeEvent( QResizeEvent * event )
 {
 	repaint();//less flickering, but resize is less responsive
-#if (VTK_MAJOR_VERSION >= 8 && defined(VTK_OPENGL2_BACKEND) )
-	QVTKOpenGLWidget::resizeEvent(event);
-#else
-	QVTKWidget::resizeEvent(event);
-#endif
+	iAVtkOldWidget::resizeEvent(event);
 }
