+++ modules/PorosityAnalyser/iAPCView.h
@@ -21,19 +21,14 @@
 #pragma once
 
 #include "ui_PCView.h"
+
+#include "iAVtkWidgetFwd.h"
 #include "qthelper/iAQTtoUIConnector.h"
 
 #include <vtkSmartPointer.h>
-#include <vtkVersion.h>
 
 typedef iAQTtoUIConnector<QDockWidget, Ui_PCView> PCViewConnector;
 
-#include <QtGlobal>
-#if (VTK_MAJOR_VERSION >= 8 && defined(VTK_OPENGL2_BACKEND) && QT_VERSION >= 0x050400 )
-class QVTKOpenGLWidget;
-#else
-class QVTKWidget;
-#endif
 class vtkContextView;
 class vtkChartParallelCoordinates;
 
@@ -56,9 +51,5 @@ protected:
 protected:
 	vtkSmartPointer<vtkContextView> m_view;
 	vtkSmartPointer<vtkChartParallelCoordinates> m_chart;
-#if (VTK_MAJOR_VERSION >= 8 && defined(VTK_OPENGL2_BACKEND) && QT_VERSION >= 0x050400 )
-	QVTKOpenGLWidget * m_widget;
-#else
-	QVTKWidget * m_widget;
-#endif
+	iAVtkOldWidget * m_widget;
 };
