+++ modules/RandomWalker/iAImageGraphTypes.h
@@ -19,12 +19,13 @@
 *          Stelzhamerstraße 23, 4600 Wels / Austria, Email: c.heinzl@fh-wels.at       *
 * ************************************************************************************/
 #pragma once
-
-#include <utility> // for std::pair
+#include "iAImageCoordinate.h"	// for iAVoxelIndexType
 
 #include <QVector>
 
-#include "iAImageCoordinate.h"	// for iAVoxelIndexType
+#include <utility> // for std::pair
+
+
 
 typedef unsigned int iAEdgeIndexType;
 typedef unsigned int iAVertexIndexType;
