### 2019-01-04 83372e5
Unify includes, headers, pragma once, spacing
