+++ modules/Segmentation/iAWatershedSegmentation.h
@@ -20,7 +20,7 @@
 * ************************************************************************************/
 #pragma once
 
-#include "iAFilter.h"
+#include <iAFilter.h>
 
 IAFILTER_DEFAULT_CLASS(iAWatershed);
 IAFILTER_DEFAULT_CLASS(iAMorphologicalWatershed);
