### 2019-01-07 fd07675
Fix build, code formatting
