+++ modules/FuzzyFeatureTracking/iAFeatureTracking.cpp
@@ -20,31 +20,37 @@
 * ************************************************************************************/
 #include "iAFeatureTracking.h"
 
+#include <vtkTable.h>
 #include <vtkTypeUInt32Array.h>
+#include <vtkVariantArray.h>
 
 #include <sstream>
 
 #define VTK_CREATE(type,name) \
   vtkSmartPointer<type> name = vtkSmartPointer<type>::New()
-	
-std::vector<std::string> &iAFeatureTracking::split(const std::string &s, char delim, std::vector<std::string> &elems) {
+
+std::vector<std::string> &iAFeatureTracking::split(const std::string &s, char delim, std::vector<std::string> &elems)
+{
 	std::stringstream ss(s);
 	std::string item;
-	while (std::getline(ss, item, delim)) {
+	while (std::getline(ss, item, delim))
+	{
 		elems.push_back(item);
 	}
 	return elems;
 }
 
-std::vector<std::string> iAFeatureTracking::split(const std::string &s, char delim) {
+std::vector<std::string> iAFeatureTracking::split(const std::string &s, char delim)
+{
 	std::vector<std::string> elems;
 	split(s, delim, elems);
 	return elems;
 }
 
-vtkTable &iAFeatureTracking::readTableFromFile(const std::string &filename, int dataLineOffset) {
-	vtkTable *t = vtkTable::New();
-	
+vtkSmartPointer<vtkTable> iAFeatureTracking::readTableFromFile(const std::string &filename, int dataLineOffset)
+{
+	auto t = vtkSmartPointer<vtkTable>::New();
+
 	VTK_CREATE(vtkTypeUInt32Array, idArray);
 	idArray->SetName("id");
 	VTK_CREATE(vtkTypeUInt32Array, posXArray);
@@ -62,18 +68,22 @@ vtkTable &iAFeatureTracking::readTableFromFile(const std::string &filename, int 
 	VTK_CREATE(vtkTypeUInt32Array, dimZVxArray);
 	dimZVxArray->SetName("dimZ");
 
-	if(filename != "") {
+	if(filename != "")
+	{
 		std::string line;
 		ifstream inputStream(filename);
-		if(inputStream.is_open()) {
-			for(int i = 0; i < dataLineOffset; i++) 
+		if(inputStream.is_open())
+		{
+			for(int i = 0; i < dataLineOffset; i++)
 				std::getline(inputStream, line);
 			// at header row
 			std::getline(inputStream, line);
 			std::vector<std::string> splittedRow;
-			while(getline(inputStream, line)) {
+			while(getline(inputStream, line))
+			{
 				splittedRow = split(line, ',');
-				try {
+				try
+				{
 					idArray->InsertNextValue(stoi(splittedRow.at(0)));
 					posXArray->InsertNextValue(stoi(splittedRow.at(5)));
 					posYArray->InsertNextValue(stoi(splittedRow.at(6)));
@@ -82,7 +92,8 @@ vtkTable &iAFeatureTracking::readTableFromFile(const std::string &filename, int 
 					dimXVxArray->InsertNextValue(stoi(splittedRow.at(10)));
 					dimYVxArray->InsertNextValue(stoi(splittedRow.at(11)));
 					dimZVxArray->InsertNextValue(stoi(splittedRow.at(12)));
-				} catch(std::exception) {}
+				}
+				catch(std::exception) {}
 			}
 		}
 	}
@@ -95,14 +106,18 @@ vtkTable &iAFeatureTracking::readTableFromFile(const std::string &filename, int 
 	t->AddColumn(dimYVxArray);
 	t->AddColumn(dimZVxArray);
 
-	return *t;
+	return t;
 }
 
[…]
