### 2019-01-07 1b419e7
FeatureCharacteristics: Migrate to new filter framework
