### 2019-01-07 acd9139
Core: QCustomPlot - OpenGL rendering fix
