### 2019-01-08 d370eb4
.toStdString/.toLatin1 fixes (mainly for special characters, e.g. umlauts, in filenames)
