+++ core/src/io/iAIO.cpp
@@ -28,6 +28,7 @@
 #include "iAConsole.h"
 #include "iAExceptionThrowingErrorObserver.h"
 #include "iAExtendedTypedCallHelper.h"
+#include "iAFileUtils.h"
 #include "iAModalityList.h"
 #include "iAOIFReader.h"
 #include "iAProgress.h"
@@ -91,7 +92,7 @@ void read_raw_image_template (unsigned long headerSize,
 {
 	typedef itk::RawImageIO<T, DIM> RawImageIOType;
 	auto io = RawImageIOType::New();
-	io->SetFileName(fileName.toLatin1().data());
+	io->SetFileName(getLocalEncodingFileName(fileName).c_str());
 	io->SetHeaderSize(headerSize);
 	for(int i=0; i<DIM; i++)
 	{
@@ -108,7 +109,7 @@ void read_raw_image_template (unsigned long headerSize,
 	typedef itk::Image< T, DIM>   InputImageType;
 	typedef itk::ImageFileReader<InputImageType> ReaderType;
 	auto reader = ReaderType::New();
-	reader->SetFileName(fileName.toLatin1().data());
+	reader->SetFileName(getLocalEncodingFileName(fileName).c_str());
 	reader->SetImageIO(io);
 	progress->Observe( reader );
 	reader->Modified();
@@ -125,7 +126,7 @@ void read_image_template(QString const & fileName, iAProgress* progress, iAConne
 	typedef itk::Image< T, DIM>   InputImageType;
 	typedef itk::ImageFileReader<InputImageType> ReaderType;
 	auto reader = ReaderType::New();
-	reader->SetFileName(fileName.toLatin1().data());
+	reader->SetFileName( getLocalEncodingFileName(fileName) );
 	progress->Observe( reader );
 	reader->Update();
 	image->SetImage(reader->GetOutput());
@@ -141,7 +142,7 @@ void write_image_template(bool compression, QString const & fileName,
 	typedef itk::Image< T, DIM>   InputImageType;
 	typedef itk::ImageFileWriter<InputImageType> WriterType;
 	auto writer = WriterType::New();
-	writer->SetFileName(fileName.toLatin1().data());
+	writer->SetFileName(getLocalEncodingFileName(fileName).c_str());
 	writer->SetInput( dynamic_cast< InputImageType * > ( image->GetITKImage() ) );
 	writer->SetUseCompression(compression);
 	progress->Observe( writer );
@@ -317,7 +318,7 @@ void iAIO::readHDF5File()
 	{
 		throw std::runtime_error("HDF5 file: Insufficient path length.");
 	}
-	hid_t file = H5Fopen(fileName.toStdString().c_str(), H5F_ACC_RDONLY, H5P_DEFAULT);
+	hid_t file = H5Fopen( getLocalEncodingFileName(fileName).c_str(), H5F_ACC_RDONLY, H5P_DEFAULT);
 	m_hdf5Path.removeLast();
 	hid_t loc_id = file;
 	QStack<hid_t> openGroups;
@@ -325,7 +326,7 @@ void iAIO::readHDF5File()
 	{
 		QString name = m_hdf5Path.last();
 		m_hdf5Path.removeLast();
-		loc_id = H5Gopen(file, name.toStdString().c_str(), H5P_DEFAULT);
+		loc_id = H5Gopen(file, name.toStdString().c_str(), H5P_DEFAULT);  // TODO: check which encoding HDF5 internal strings have!
 		openGroups.push(loc_id);
 	}
 
@@ -480,7 +481,7 @@ void iAIO::run()
 				// TODO: write more than one modality!
 				auto img = getVtkImageData();
[…]
