### 2019-01-08 bd0f99f
Core: Improve open with datatype conversion
