+++ core/src/dlg_datatypeconversion.cpp
@@ -20,7 +20,8 @@
 * ************************************************************************************/
 #include "dlg_datatypeconversion.h"
 
-#include "charts/iAHistogramWidget.h"
+#include "charts/iAHistogramData.h"
+#include "charts/iAPlotTypes.h"
 #include "iAConnector.h"
 #include "iAToolsITK.h"
 #include "iAToolsVTK.h"
@@ -74,7 +75,7 @@ namespace
 	}
 }
 
-template<class T> void DataTypeConversion_template(QString const & filename, double* b, iAPlotData::DataType * histptr, float* m_min, float* m_max, float* m_dis, iAConnector* xyconvertimage, iAConnector* xzconvertimage, iAConnector* yzconvertimage)
+template<class T> void DataTypeConversion_template(QString const & filename, double* b, iAPlotData::DataType * histptr, float* m_min, float* m_max, float* m_dis, iAConnector* xyconvertimage, iAConnector* xzconvertimage/*, iAConnector* yzconvertimage*/)
 {
 	// TODO: use itk methods instead?
 	typedef itk::Image< T, 3 >   InputImageType;
@@ -109,9 +110,9 @@ template<class T> void DataTypeConversion_template(QString const & filename, dou
 	while ( (result = fread (reinterpret_cast<char*>(&buffer),datatypesize, elemCount, pFile)) == elemCount )
 	{
 		if (buffer < min)
-		{	min = buffer;	}
+			min = buffer;
 		if (buffer > max)
-		{	max = buffer;	}
+			max = buffer;
 	}
 
 	float discretization = (float)((max-min)/(b[7]));
@@ -155,7 +156,10 @@ template<class T> void DataTypeConversion_template(QString const & filename, dou
 	typename TwoDInputImageType::IndexType extractindex; extractindex.Fill(0);	extractregion.SetIndex(extractindex);
 	typename TwoDInputImageType::PointType extractpoint; extractpoint.Fill(0);
 	typename TwoDInputImageType::SpacingType extractspacing; extractspacing[0] = b[4];	extractspacing[1] = b[5];	extractspacing[2] = b[6];
-	// along z axis - xy plane
+
+
+	// XY plane - along z axis
+
 	typename TwoDInputImageType::SizeType extractsize;	extractsize[0] = b[1]; extractsize[1] = b[2];	extractsize[2] = 1;
 	extractregion.SetSize(extractsize);
 
@@ -211,13 +215,16 @@ template<class T> void DataTypeConversion_template(QString const & filename, dou
 	xyconvertimage->SetImage(xyrescalefilter->GetOutput());
 	xyconvertimage->Modified();
 
-	vtkSmartPointer<vtkMetaImageWriter> metaImageWriter = vtkSmartPointer<vtkMetaImageWriter>::New();
-	metaImageWriter->SetFileName("xyimage.mhd");
-	metaImageWriter->SetInputData(xyconvertimage->GetVTKImage());
-	metaImageWriter->SetCompression(0);
-	metaImageWriter->Write();
+	// DEBUG:
+	//vtkSmartPointer<vtkMetaImageWriter> metaImageWriter = vtkSmartPointer<vtkMetaImageWriter>::New();
+	//metaImageWriter->SetFileName("xyimage.mhd");
+	//metaImageWriter->SetInputData(xyconvertimage->GetVTKImage());
+	//metaImageWriter->SetCompression(0);
+	//metaImageWriter->Write();
 
-	//xz plane - along y axis
+	
[…]
