### 2019-01-08 82c4ce8
Core: Open with datatype conversion improvements
