+++ core/src/dlg_datatypeconversion.cpp
@@ -20,6 +20,7 @@
 * ************************************************************************************/
 #include "dlg_datatypeconversion.h"
 
+#include "charts/iAChartWidget.h"
 #include "charts/iAHistogramData.h"
 #include "charts/iAPlotTypes.h"
 #include "iAConnector.h"
@@ -37,21 +38,17 @@
 
 #include <vtkColorTransferFunction.h>
 #include <vtkGenericOpenGLRenderWindow.h>
-#include <vtkImageAccumulate.h>
 #include <vtkImageActor.h>
 #include <vtkImageData.h>
+#include <vtkImageMapper3D.h>
 #include <vtkImageMapToColors.h>
-#include <vtkImageReslice.h>
 #include <vtkInteractorStyleImage.h>
-#include <vtkLookupTable.h>
 #include <vtkMatrix4x4.h>
 #include <vtkMetaImageWriter.h>
 #include <vtkPlaneSource.h>
-#include <vtkPiecewiseFunction.h>
 #include <vtkPolyDataMapper.h>
 #include <vtkProperty.h>
 #include <vtkRenderer.h>
-#include <vtkRenderWindowInteractor.h>
 
 #include <QCheckBox>
 #include <QComboBox>
@@ -75,7 +72,25 @@ namespace
 	}
 }
 
-template<class T> void DataTypeConversion_template(QString const & filename, double* b, iAPlotData::DataType * histptr, float* m_min, float* m_max, float* m_dis, iAConnector* xyconvertimage, iAConnector* xzconvertimage/*, iAConnector* yzconvertimage*/)
+template<class T> void getFileMinMax(FILE* pFile, double& minVal, double& maxVal)
+{
+	minVal = std::numeric_limits<double>::max();
+	maxVal = std::numeric_limits<double>::lowest();
+	//calculation of minimum and maximum
+	fseek(pFile, 0, SEEK_SET);
+	const int elemCount = 1;
+	T buffer;
+	size_t result;
+	while ((result = fread(reinterpret_cast<char*>(&buffer), sizeof(buffer), elemCount, pFile)) == elemCount)
+	{
+		if (buffer < minVal)
+			minVal = buffer;
+		if (buffer > maxVal)
+			maxVal = buffer;
+	}
+}
+
+template<class T> void DataTypeConversion_template(QString const & filename, double* b, iAPlotData::DataType * histptr, double & minVal, double & maxVal, double & discretization, iAConnector* xyimage, iAConnector* xzimage, iAConnector* yzimage)
 {
 	// TODO: use itk methods instead?
 	typedef itk::Image< T, 3 >   InputImageType;
@@ -97,38 +112,26 @@ template<class T> void DataTypeConversion_template(QString const & filename, dou
 	unsigned long datatypesize = sizeof(buffer);
 	long slicesize = b[1]*b[2];
 	int slicecounter = 0;
-	size_t result;
 	int numsliceread = 0;
-
 	long totalsize = b[1]*b[2]*b[3]*datatypesize;
-	double min = std::numeric_limits<double>::max();
-	double max = std::numeric_limits<double>::lowest();
 
-	//calculation of minimum and maximum
-	fseek ( pFile , 0 , SEEK_SET );
-	const int elemCount = 1;
-	while ( (result = fread (reinterpret_cast<char*>(&buffer),datatypesize, elemCount, pFile)) == elemCount )
-	{
-		if (buffer < min)
-			min = buffer;
-		if (buffer > max)
-			max = buffer;
-	}
-
[…]
