### 2019-01-08 afff156
Core: Open with datatype conversion fixes
