### 2019-01-08 6c32da0
Core: Open with datatype conversion: remove duplicate code
