+++ core/src/dlg_datatypeconversion.cpp
@@ -90,6 +90,92 @@ template<class T> void getFileMinMax(FILE* pFile, double& minVal, double& maxVal
 	}
 }
 
+template <class T> void extractSliceImage(typename itk::Image<T, 3>::Pointer itkimage, int firstDir, int secondDir, /*int sliceNr, projectionMethod[=SUM], */iAConnector* image)
+{
+	typedef typename itk::Image<T, 3> InputImageType;
+	typedef typename itk::Image<double, 3> TwoDInputImageType;
+
+	typename TwoDInputImageType::RegionType extractregion;
+	typename TwoDInputImageType::IndexType extractindex; extractindex.Fill(0);	extractregion.SetIndex(extractindex);
+	typename TwoDInputImageType::PointType extractpoint; extractpoint.Fill(0);
+	typename TwoDInputImageType::SpacingType extractspacing;
+	extractspacing[0] = itkimage->GetSpacing()[firstDir];
+	extractspacing[1] = itkimage->GetSpacing()[secondDir];
+	extractspacing[2] = 1;
+
+	typename TwoDInputImageType::SizeType extractsize;
+	extractsize[0] = itkimage->GetLargestPossibleRegion().GetSize()[firstDir];
+	extractsize[1] = itkimage->GetLargestPossibleRegion().GetSize()[secondDir];
+	extractsize[2] = 1;
+	extractregion.SetSize(extractsize);
+
+	typename TwoDInputImageType::Pointer twodimage = TwoDInputImageType::New();
+	twodimage->SetRegions(extractregion);
+	twodimage->SetSpacing(extractspacing);
+	twodimage->SetOrigin(extractpoint);
+	twodimage->Allocate();
+	twodimage->FillBuffer(0);
+
+	typedef itk::ImageRegionIterator<TwoDInputImageType> twoditeratortype;
+	twoditeratortype iter(twodimage, twodimage->GetRequestedRegion());
+
+	//create slice iterator
+	typedef itk::ImageSliceConstIteratorWithIndex<InputImageType> SliceIteratorType;
+	SliceIteratorType SliceIter(itkimage, itkimage->GetRequestedRegion());
+
+	//set direction
+	SliceIter.SetFirstDirection(firstDir);
+	SliceIter.SetSecondDirection(secondDir);
+
+	SliceIter.GoToBegin();
+	//int curSlice = 0;
+	while (!SliceIter.IsAtEnd())
+	{
+		//if (curSlice == sliceNr)
+		//{
+			iter.GoToBegin();
+			while (!SliceIter.IsAtEndOfSlice())
+			{
+				while (!SliceIter.IsAtEndOfLine())
+				{
+					typename InputImageType::PixelType value = SliceIter.Get();
+					typename InputImageType::IndexType index = SliceIter.GetIndex();
+					typename TwoDInputImageType::PixelType pix = iter.Get() + value;  // sum projection
+					iter.Set(pix);
+					++iter;
+					++SliceIter;
+				}
+				SliceIter.NextLine();
+			}
+		//	break;
+		//}
+		++curSlice;
+		SliceIter.NextSlice();
+	}
+
+	typedef itk::NormalizeImageFilter<TwoDInputImageType, TwoDInputImageType> NIFTYpe;
+	typename NIFTYpe::Pointer normalizefilter = NIFTYpe::New();
+	normalizefilter->SetInput(twodimage);
+	normalizefilter->Update();
+
+	typedef itk::RescaleIntensityImageFilter<TwoDInputImageType, TwoDInputImageType> RIIFType;
[…]
