### 2019-01-08 0252e2d
Linux build fixes
