+++ core/src/io/iAOIFReader.cpp
@@ -38,7 +38,8 @@ void Reader::read(QString const & filename, iAConnector* con, int channel,
 	std::vector<vtkSmartPointer<vtkImageData> > * volumes)
 {
 	OIFReader* reader = new OIFReader();
-	reader->SetFile(filename.toStdWString());
+	auto wfn = filename.toStdWString();
+	reader->SetFile(wfn);
 	reader->SetSliceSeq(false);
 	std::wstring timeId(L"_T");
 	reader->SetTimeId(timeId);
