### 2019-01-09 52277db
CMake: Remove unreachable FindCUDA code
