### 2019-01-09 fb5275c
Core: Don't use tr for storing settings
