+++ core/src/mainwindow.cpp
@@ -899,21 +899,21 @@ void MainWindow::saveRenderSettings(QDomDocument &doc)
 
 	// add new camera node
 	QDomElement renderSettingsElement = doc.createElement("renderSettings");
-	renderSettingsElement.setAttribute("showSlicers", tr("%1").arg(defaultRenderSettings.ShowSlicers));
-	renderSettingsElement.setAttribute("showSlicePlanes", tr("%1").arg(defaultRenderSettings.ShowSlicePlanes));
-	renderSettingsElement.setAttribute("showHelpers", tr("%1").arg(defaultRenderSettings.ShowHelpers));
-	renderSettingsElement.setAttribute("showRPosition", tr("%1").arg(defaultRenderSettings.ShowRPosition));
-	renderSettingsElement.setAttribute("parallelProjection", tr("%1").arg(defaultRenderSettings.ParallelProjection));
-	renderSettingsElement.setAttribute("backgroundTop", tr("%1").arg(defaultRenderSettings.BackgroundTop.toLatin1().constData()));
-	renderSettingsElement.setAttribute("backgroundBottom", tr("%1").arg(defaultRenderSettings.BackgroundBottom.toLatin1().constData()));
-	renderSettingsElement.setAttribute("linearInterpolation", tr("%1").arg(defaultVolumeSettings.LinearInterpolation));
-	renderSettingsElement.setAttribute("shading", tr("%1").arg(defaultVolumeSettings.Shading));
-	renderSettingsElement.setAttribute("sampleDistance", tr("%1").arg(defaultVolumeSettings.SampleDistance));
-	renderSettingsElement.setAttribute("ambientLighting", tr("%1").arg(defaultVolumeSettings.AmbientLighting));
-	renderSettingsElement.setAttribute("diffuseLighting", tr("%1").arg(defaultVolumeSettings.DiffuseLighting));
-	renderSettingsElement.setAttribute("specularLighting", tr("%1").arg(defaultVolumeSettings.SpecularLighting));
-	renderSettingsElement.setAttribute("specularPower", tr("%1").arg(defaultVolumeSettings.SpecularPower));
-	renderSettingsElement.setAttribute("renderMode", tr("%1").arg(defaultVolumeSettings.RenderMode));
+	renderSettingsElement.setAttribute("showSlicers", defaultRenderSettings.ShowSlicers);
+	renderSettingsElement.setAttribute("showSlicePlanes", defaultRenderSettings.ShowSlicePlanes);
+	renderSettingsElement.setAttribute("showHelpers", defaultRenderSettings.ShowHelpers);
+	renderSettingsElement.setAttribute("showRPosition",defaultRenderSettings.ShowRPosition);
+	renderSettingsElement.setAttribute("parallelProjection", defaultRenderSettings.ParallelProjection);
+	renderSettingsElement.setAttribute("backgroundTop", defaultRenderSettings.BackgroundTop);
+	renderSettingsElement.setAttribute("backgroundBottom", defaultRenderSettings.BackgroundBottom);
+	renderSettingsElement.setAttribute("linearInterpolation", defaultVolumeSettings.LinearInterpolation);
+	renderSettingsElement.setAttribute("shading", defaultVolumeSettings.Shading);
+	renderSettingsElement.setAttribute("sampleDistance", defaultVolumeSettings.SampleDistance);
+	renderSettingsElement.setAttribute("ambientLighting", defaultVolumeSettings.AmbientLighting);
[…]
