### 2019-01-09 d010df8
Core: Fix slicer image stack saving
