+++ core/src/io/iAAmiraMeshIO.cpp
@@ -25,8 +25,10 @@
 
 #include <vtkImageData.h>
 
+#include <QFile>
 #include <QString>
 #include <QStringList>
+#include <QTextStream>
 
 #include <cstdio>
 #include <cstring>
@@ -100,35 +102,29 @@ vtkSmartPointer<vtkImageData> iAAmiraMeshIO::Load(QString const & fileName)
 	FILE* fp = fopen( getLocalEncodingFileName(fileName).c_str(), "rb");
 	if (!fp)
 	{
-		DEBUG_LOG(QString("Could not open file '%1'.").arg(fileName));
-		return vtkSmartPointer<vtkImageData>();
+		throw std::runtime_error(QString("Could not open file '%1'.").arg(fileName).toStdString());
 	}
-
-	//DEBUG_LOG(QString("Reading %1").arg(fileName));
 
 	//We read the first 2k bytes into memory to parse the header.
 	//The fixed buffer size looks a bit like a hack, and it is one, but it gets the job done.
-	// TODO: use proper parser!
-	const size_t MAX_HEADER_SIZE = 2047;
+	const size_t MaxHeaderSize = 2047;
+	char buffer[MaxHeaderSize +1];
+	size_t readBytes = fread(buffer, sizeof(char), MaxHeaderSize, fp);
 
-	char buffer[MAX_HEADER_SIZE+1];
-	size_t readBytes = fread(buffer, sizeof(char), MAX_HEADER_SIZE, fp);
-	if (readBytes != MAX_HEADER_SIZE)
+	if (readBytes == 0 || ferror(fp) != 0)
 	{
-		DEBUG_LOG(QString("Could not read first %1 bytes of Avizo/AmiraMesh file %2.").arg(MAX_HEADER_SIZE).arg(fileName));
 		fclose(fp);
-		return vtkSmartPointer<vtkImageData>();
+		throw std::runtime_error(QString("Could not read header of Avizo/AmiraMesh file %1.").arg(MaxHeaderSize).arg(fileName).toStdString());
 	}
-	buffer[MAX_HEADER_SIZE-1] = '\0'; //The following string routines prefer null-terminated strings
+	buffer[readBytes-1] = '\0'; //The following string routines prefer null-terminated strings
 
 	QString header(buffer);
 
 	if (!header.startsWith(AmiraMeshFileTag) &&
 		!header.startsWith(AvizoFileTag))
 	{
-		DEBUG_LOG(QString("Not a proper Avizo/AmiraMesh file: %1.").arg(fileName));
 		fclose(fp);
-		return vtkSmartPointer<vtkImageData>();
+		throw std::runtime_error(QString("File %1 is not a proper Avizo/AmiraMesh file, it is missing the initial file tag.").arg(fileName).toStdString());
 	}
 
 	//Find the Lattice definition, i.e., the dimensions of the uniform grid
@@ -172,12 +168,13 @@ vtkSmartPointer<vtkImageData> iAAmiraMeshIO::Load(QString const & fileName)
 	else if (dataTypeStr.startsWith(FloatType))
 	{
 		//A field with more than one component, i.e., a vector field
+		dataType = VTK_FLOAT;
 		sscanf(FindAndJump(buffer, "Lattice { float["), "%d", &NumComponents);
 	}
 	else
 	{
-		DEBUG_LOG(QString("Unknown pixel type '%1' (not yet implemented).").arg(dataTypeStr));
-		return vtkSmartPointer<vtkImageData>();
+		fclose(fp);
+		throw std::runtime_error(QString("Unknown pixel type '%1' (not yet implemented). Supported pixel types: byte (unsigned char), float.").arg(dataTypeStr).toStdString());
 	}
[…]
