### 2019-01-09 35c9c66
Core: Disable non-functional NRRD file format loading
