### 2019-01-09 5f54f55
Core: Remove orientation/rotation fix comments
