### 2019-01-09 83cd8fd
Minor refactoring
