### 2019-01-09 329b4fa
Core: Fix .stl file saving (if no volume loaded)
