### 2019-01-09 8d0b7bb
Core: prevent modality file name change if storing .stl
