### 2019-01-09 d0babab
Batch: Stack reader - abort for unknown file extension
