### 2019-01-09 8719ccf
Connected Components: FileNameSave type for Label file parameter, minor description fixes
