### 2019-01-09 80d2f65
FeatureCharacteristics: Default empty csv filename instead of 0
