### 2019-01-09 f22e19f
FeatureScout: Minor fix MObject STL file saving
