### 2019-01-10 cf20e78
Core: Fix volume stack loading
