### 2019-01-10 8aa8401
FuzzyFeatureTracking: Improved error checking/message, fix qrc ref.
