### 2019-01-10 fadde1b
PorosityAnalyser: Minor refactoring
