+++ modules/PorosityAnalyser/Dependencies.cmake
@@ -1,4 +1,3 @@
-# iAnalyseToolkit directories
 SET (DEPENDENCIES_IA_TOOLKIT_DIRS
 	MaximumDistance
 )
