### 2019-01-11 af10bfb
GEMSe: Set UTF-8 encoding on QSettings/QTextStream
