+++ modules/GEMSe/iASamplingResults.cpp
@@ -90,6 +90,7 @@ QSharedPointer<iASamplingResults> iASamplingResults::Load(QString const & smpFil
 	}
 	// TODO: replace with QSettings?
 	QTextStream in(&file);
+	in.setCodec("UTF-8");
 	QFileInfo fileInfo(file);
 	if (in.atEnd())
 	{
@@ -149,6 +150,7 @@ bool iASamplingResults::Store(QString const & fileName,
 		return false;
 	}
 	QTextStream out(&paramRangeFile);
+	out.setCodec("UTF-8");
 	QFileInfo fi(paramRangeFile);
 	out << SMPFileFormatVersion << endl;
 	out << "Name" << Output::NameSeparator << m_name << endl;
