### 2019-01-11 5437642
FuzzyFeatureTracking: Remove blocking interactor->start()
