### 2019-01-12 #1139 @katyhuff
Hi @codeling, thanks for this submission. I'm sorry for the delay. I'm back in action now. Now, @codeling , let's ping a few fitting reviewers, with C++ and Python expertise as well as some background in image analysis and/or computed tomography. We can start the review once two or three of them accept the request to review.

Dear potential reviewers (@krischer @rrrlw @camillescott  ),

Based on your expertise in C++ and Python, image analysis, and/or computed tomography, I would like to ask if you can help us out with a submission to JOSS (The Journal of Open Source Software), where I’m an editor.

Title: open_iA: A tool for processing and visual analysis of industrial computed tomography datasets
Summary: Visual analysis and processing of volumetric datasets, with a focus on industrial computed tomography
Wiki: https://github.com/3dct/open_ia/wiki
Topic: Data visualization, signal processing, imaging
Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01139/joss.01139/10.21105.joss.01139.pdf
Submitting author: @codeling (Bernhard Fröhler)
Repository: https://github.com/3dct/open_iA
Version: 2018.12

@krischer : Lion, are you able to and interested in taking on this review?
@rrrlw : Raoul, are you able to and interested in taking on this review?
@camillescott : Camille, are you able to and interested in taking on this review?

Your prompt response is appreciated. Please see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html for more information about reviewer criteria and the process if it is new to you.
