### 2019-01-14 #1139 @katyhuff
Thank you for letting us know, @rrrlw .

@codeling : Can you suggest any reviewers? It would be helpful to me if you can think of anyone. The folks on [this list](https://bit.ly/joss-reviewers) are already reviewers for JOSS. If there is someone else that you would like to suggest, please let me know. Though they will need to be free of any conflict of interest, you may be aware of individuals in your community who would be good reviewers for this submission.
