### 2019-01-14 18bae31
GEMSe: Minor fixes
