### 2019-01-14 7785ab0
GEMSe: Chart marker (potential crash) fixes
