### 2019-01-14 31885f2
FeatureScout: CSV export in CSV input dialog
