### 2019-01-14 #1139 @katyhuff
Both would be excellent if they have the time, indeed.

Pawel (@jochym) and Tania (@trallard), based on your expertise, I would like to ask if you can help us out by volunteering to review this submission to JOSS.

Title: open_iA: A tool for processing and visual analysis of industrial computed tomography datasets
Summary: Visual analysis and processing of volumetric datasets, with a focus on industrial computed tomography
Wiki: https://github.com/3dct/open_ia/wiki
Topic: Data visualization, signal processing, imaging
Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01139/joss.01139/10.21105.joss.01139.pdf
Submitting author: @codeling (Bernhard Fröhler)
Repository: https://github.com/3dct/open_iA
Version: 2018.12
