### 2019-01-15 fcbe6b5
XRF: Remove superfluous include
