### 2019-01-16 9cf0d14
PorosityAnalyzer: additional error checks
