### 2019-01-16 e2ace3a
FIAKEr: Make work with less than 25 fibers
