### 2019-01-17 82c4456
FIAKER: Load and store project files in UTF-8 encoding
