### 2019-01-17 f44ad46
FIAKER: Remove fixed 74.5 optimization step coordinate shift, show bounding box information
