### 2019-01-17 #1139 @katyhuff
@behollister and @ngoldbaum , would either of you be available to and interested in reviewing this JOSS submission? I agree that you like have the expertise. Do you have the time? We'd like at least two reviewers. @trallard has thankfully already agreed, so we're looking for one or two more.

Title: open_iA: A tool for processing and visual analysis of industrial computed tomography datasets
Summary: Visual analysis and processing of volumetric datasets, with a focus on industrial computed tomography
Wiki: https://github.com/3dct/open_ia/wiki
Topic: Data visualization, signal processing, imaging
Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01139/joss.01139/10.21105.joss.01139.pdf
Submitting author: @codeling (Bernhard Fröhler)
Repository: https://github.com/3dct/open_iA
Version: 2018.12
