### 2019-01-17 #1139 @behollister
Hi,

Yes, I can review the submission. When is the deadline?

Regards,

Brad

Brad E. Hollister, Ph.D.

Department of Computer Science
Californian State University Dominguez Hills
Office: NSM B218
Phone: 864-395-5560
