### 2019-01-17 #1139 @katyhuff
Thaknks @behollister ! We aim to complete the review process quickly, so I typically expect a reviewer to get started with their review during the week following assignment, complete their initial review within a week or so thereafter, and remain active with the back-and-forth until the paper is accepted (can be as quick as a week, but is often 3 weeks, and occasionally longer.) If a deadline is helpful to you, I can certainly assign one, though I typically do not.

Please see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html for more information about reviewer criteria and the process.
