### 2019-01-18 4116fe4
Enabled class export with cylinder visualisation
