### 2019-01-18 f32be87
ExportClass switch between single and multiclass output
