+++ modules/FeatureScout/dlg_FeatureScout.cpp
@@ -1985,31 +1985,45 @@ void dlg_FeatureScout::CreateLabelledOutputMask(iAConnector & con, const QString
 	typedef itk::Image<ClassIDType, DIM>   OutputImageType;
 	OutputImageType::SizeType OutputImageSize;
 
-	bool fiberIDLabelling = (classTreeModel->invisibleRootItem()->rowCount() >= 2 &&
-		(m_renderMode == rmSingleClass && activeClassItem->row() > 0));
-	if (fiberIDLabelling &&
-		(QMessageBox::question(this, "FeatureScout", "Only one class selected, "
-			"should we export the individual fiber IDs? "
-			"If you select No, all fibers in the class will be labelled with the class ID.",
-			QMessageBox::Yes | QMessageBox::No)
-			== QMessageBox::No))
-	{
-		fiberIDLabelling = false;
-	}
-
 	QMap<size_t, ClassIDType> currentEntries;
+
+	QString mode1 = "Export all classes";
+	QString mode2 = "Export single selected class";
+	QStringList modes = (QStringList() << tr(mode1.toStdString().c_str()) << tr(mode2.toStdString().c_str()));
+	QStringList inList = (QStringList() << tr("+Classification"));
+	QList<QVariant> inPara = (QList<QVariant>() << modes);
+	dlg_commoninput dlg(this, "Save classification options", inList, inPara, NULL);
+	if (dlg.exec() != QDialog::Accepted)
+		return;
+	QString mode = dlg.getComboBoxValue(0);
+	bool exportAllClassified = (mode.compare(mode1) == 0); //if export all selected else single class export
+
 	// Skip first, as classes start with 1, 0 is the uncategorized class
 	for (int i = 1; i < classTreeModel->invisibleRootItem()->rowCount(); i++)
 	{
-		if (m_renderMode == rmSingleClass && i != activeClassItem->row())
+	
+		if (!exportAllClassified && i != activeClassItem->row())
 			continue;
-		auto x = classTreeModel->invisibleRootItem()->rowCount();
+		
 		QStandardItem *item = classTreeModel->invisibleRootItem()->child(i);
 		for (int j = 0; j < item->rowCount(); j++)
 		{
 			size_t labelID = item->child(j)->text().toULongLong();
 			currentEntries.insert(labelID, i);
 		}
+	}
+
+	//export class ID or fiber labels for single class
+	bool fiberIDLabelling = (classTreeModel->invisibleRootItem()->rowCount() >= 2 &&
+		(m_renderMode == rmSingleClass && activeClassItem->row() > 0) && (!exportAllClassified ));
+	if (fiberIDLabelling &&
+		(QMessageBox::question(this, "FeatureScout", "Only one class selected, "
+			"should we export the individual fiber IDs? "
+			"If you select No, all fibers in the class will be labeled with the class ID.",
+			QMessageBox::Yes | QMessageBox::No)
+			== QMessageBox::No))
+	{
+		fiberIDLabelling = false;
 	}
 
 	auto in_img = dynamic_cast<InputImageType*>(con.GetITKImage());
