### 2019-01-21 6c78b51
FIAKER: Make cylinder diameter adaptable via slider
