### 2019-01-21 ca2bc06
FIAKER: Show bounding box for selected fiber
