### 2019-01-23 60f3b79
FIAKER: Separate step/general coord. shift, make step shift configurable in project
