+++ modules/FiAKEr/iAFiAKErController.h
@@ -79,7 +79,7 @@ class iAFiAKErController : public QMainWindow, public iASelectionProvider
 public:
 	typedef std::vector<std::vector<size_t> > SelectionType;
 	iAFiAKErController(MainWindow* mainWnd);
-	void start(QString const & path, QString const & configName);
+	void start(QString const & path, QString const & configName, double stepShift);
 	~iAFiAKErController();
 	std::vector<std::vector<size_t> > & selection() override;
 	static void loadAnalysis(MainWindow* mainWnd, QString const & folder);
