### 2019-01-24 1b5a1d3
Fix VTK < 8 build
