### 2019-01-24 547a9bf
Core: iAVec3 extensions
