### 2019-01-24 b459613
FIAKER: WIP - Context merging and spacing
