+++ modules/FiAKEr/iAFiAKErController.cpp
@@ -163,7 +163,10 @@ iAFiAKErController::iAFiAKErController(MainWindow* mainWnd) :
 	m_playTimer(new QTimer(this)),
 	m_refDistCompute(nullptr),
 	m_renderManager(new iARendererManager()),
-	m_colorByThemeName(iALUT::GetColorMapNames()[0])
+	m_colorByThemeName(iALUT::GetColorMapNames()[0]),
+	m_showFiberContext(false),
+	m_mergeContextBoxes(false),
+	m_contextSpacing(0.0)
 {
 	setDockOptions(AllowNestedDocks | AllowTabbedDocks);
 #if QT_VERSION >= QT_VERSION_CHECK(5, 6, 0)
@@ -296,13 +299,30 @@ void iAFiAKErController::resultsLoaded()
 	sliderWidgetLayout->addWidget(diameterFactorSlider, 2, 1);
 	sliderWidgetLayout->addWidget(m_diameterFactorLabel, 2, 2);
 
-	QCheckBox* fiberContextWidget = new QCheckBox("Show selected fiber context");
-	connect(fiberContextWidget, &QCheckBox::stateChanged, this, &iAFiAKErController::showFiberContextChanged);
-	/*
 	QWidget* fiberContextWidget = new QWidget();
-	auto fiberContextLayout = new QHBoxLayout();
-	fiberContextWidget
-	*/
+	auto fiberContextLayout = new QVBoxLayout();
+	fiberContextWidget->setLayout(fiberContextLayout);
+	fiberContextLayout->setContentsMargins(0, 0, 0, 0);
+	fiberContextLayout->setSpacing(SettingSpacing);
+	QCheckBox* cbShowFiberContext = new QCheckBox("Show selected fiber context");
+	connect(cbShowFiberContext, &QCheckBox::stateChanged, this, &iAFiAKErController::showFiberContextChanged);
+	fiberContextLayout->addWidget(cbShowFiberContext);
+	QCheckBox* cbMergeFiberContextBoxes = new QCheckBox("Merge fiber context boxes");
+	connect(cbMergeFiberContextBoxes, &QCheckBox::stateChanged, this, &iAFiAKErController::mergeFiberContextBoxesChanged);
+	fiberContextLayout->addWidget(cbMergeFiberContextBoxes);
+	QWidget* fiberContextSpacingWidget = new QWidget();
+	auto fiberContextSpacingLayout = new QHBoxLayout();
+	fiberContextSpacingWidget->setLayout(fiberContextSpacingLayout);
+	fiberContextSpacingLayout->setContentsMargins(0, 0, 0, 0);
+	fiberContextSpacingLayout->setSpacing(SettingSpacing);
+	QDoubleSpinBox* sbContextSpacing = new QDoubleSpinBox();
+	sbContextSpacing->setRange(0.0, 1000.0);
+	sbContextSpacing->setValue(0.0);
+	fiberContextSpacingLayout->addWidget(new QLabel("Context Spacing"));
+	fiberContextSpacingLayout->addWidget(sbContextSpacing);
+	fiberContextLayout->addWidget(fiberContextSpacingWidget);
+	connect(sbContextSpacing, SIGNAL(valueChanged(double)), this, SLOT(contextSpacingChanged(double)));
+
 
 	QWidget* moreButtons = new QWidget();
 	moreButtons->setLayout(new QHBoxLayout());
@@ -1562,43 +1582,100 @@ void iAFiAKErController::diameterFactorChanged(int diameterFactorInt)
 	}
 }
 
-void iAFiAKErController::showFiberContextChanged()
+void iAFiAKErController::showFiberContextChanged(int newState)
 {
-	for (auto actor: m_contextActors)
[…]
