### 2019-01-25 2653318
Fix module loading with dependencies
