### 2019-01-25 5c55c16
GEMSe: Better handling of infinity avg. entropy
