### 2019-01-25 ba1e72b
FIAKER: Context box update fixes
