### 2019-01-28 3dct/open_iA#32 @trallard
[JOSS REVIEW] Install instructions
Related to [https://github.com/openjournals/joss-reviews/issues/1185](https://github.com/openjournals/joss-reviews/issues/1185)

There are detailed instructions for development installs, but there are no instructions for the general usage (i.e. I want to use the tool but not necessarily add a new function).

It might be worth to:
- Package the latest release in a way that could be relatively easily installed by both Windows and unix users
- Provide the instructions on the wiki in addition to the development ones
