### 2019-01-28 47570ab
FIAKER: split GUI initializations into one method per view
