+++ modules/FiAKEr/iAFiAKErController.cpp
@@ -217,8 +217,47 @@ void iAFiAKErController::resultsLoaded()
 	m_resultUIs.resize(m_data->result.size());
 	m_selection.resize(m_data->result.size());
 
-	// Main 3D View:
+	auto main3DView = setupMain3DView();
+	auto settingsView = setupSettingsView();
+	auto optimStepsView = setupOptimStepView();
+	auto resultListView = setupResultListView();
+	auto protocolView = setupProtocolView();
+	auto selectionView = setupSelectionView();
 
+	m_views.resize(DockWidgetCount);
+	m_views[ResultListView] = new iADockWidgetWrapper(resultListView, "Result list", "foeResultList");
+	m_views[Main3DView]     = new iADockWidgetWrapper(main3DView, "3D view", "foe3DView");
+	m_views[OptimStepChart] = new iADockWidgetWrapper(optimStepsView, "Optimization Steps", "foeTimeSteps");
+	m_views[SPMView]        = new iADockWidgetWrapper(m_spm, "Scatter Plot Matrix", "foeSPM");
+	m_views[ProtocolView]   = new iADockWidgetWrapper(protocolView, "Interactions", "foeInteractions");
+	m_views[SelectionView]  = new iADockWidgetWrapper(selectionView, "Selections", "foeSelections");
+	m_views[SettingsView]   = new iADockWidgetWrapper(settingsView, "Settings", "foeSettings");
+
+	splitDockWidget(m_views[JobView], m_views[ResultListView], Qt::Vertical);
+	splitDockWidget(m_views[ResultListView], m_views[Main3DView], Qt::Horizontal);
+	splitDockWidget(m_views[ResultListView], m_views[OptimStepChart], Qt::Vertical);
+	splitDockWidget(m_views[Main3DView], m_views[SPMView], Qt::Horizontal);
+	splitDockWidget(m_views[ResultListView], m_views[ProtocolView], Qt::Vertical);
+	splitDockWidget(m_views[Main3DView], m_views[SelectionView], Qt::Vertical);
+	splitDockWidget(m_views[Main3DView], m_views[SettingsView], Qt::Vertical);
+
+	loadStateAndShow();
+}
+
+iAFiAKErController::~iAFiAKErController()
+{
+	if (parent())
+	{
+		QSettings settings;
+		settings.setValue(ModuleSettingsKey + "/maximized", isMaximized());
+		if (!isMaximized())
+			settings.setValue(ModuleSettingsKey + "/geometry", qobject_cast<QWidget*>(parent())->geometry());
+		settings.setValue(ModuleSettingsKey + "/state", saveState());
+	}
+}
+
+QWidget * iAFiAKErController::setupMain3DView()
+{
 	m_mainRenderer = new iAVtkWidgetClass();
 	auto renWin = vtkSmartPointer<vtkGenericOpenGLRenderWindow>::New();
 	auto ren = vtkSmartPointer<vtkRenderer>::New();
@@ -259,10 +298,11 @@ void iAFiAKErController::resultsLoaded()
 	mainRendererContainer->layout()->setContentsMargins(DockWidgetMargin, DockWidgetMargin, DockWidgetMargin, DockWidgetMargin);
 	mainRendererContainer->layout()->addWidget(m_mainRenderer);
 	mainRendererContainer->layout()->addWidget(m_showReferenceWidget);
+	return mainRendererContainer;
+}
 
-
-	// Settings View
-
+QWidget* iAFiAKErController::setupSettingsView()
+{
 	m_defaultOpacitySlider = new QSlider(Qt::Horizontal);
 	m_defaultOpacitySlider->setMinimum(0);
 	m_defaultOpacitySlider->setMaximum(255);
[…]
