### 2019-01-28 3dct/open_iA#32 @codeling
Thanks for the feedback!

- We try to provide installable binaries for testing/using our software through the [open_iA releases](https://github.com/3dct/open_iA/releases) page; we currently provide binaries for Windows (7-10) as well as Linux. The linux build should work at least on current Ubuntu and derivatives; we unfortunately don't know an easy mechanism at the moment for providing binaries that work across multiple linux/unix distributions, and we don't have the resources to run builds for multiple distributions. If you have any proposals how we could improve that situation we'd be very glad! We have started looking into the flatpak format, and might provide such packages in the near future, but so far we couldn't get our program to reliably run within its sandboxing mechanisms, the major issue was with access to files.

- As for instructions, our Readme.md references the releases page ("For trying it out, download the latest binary version from github."), as well as the wiki front page (section "Getting open_iA"). Do you have any suggestions how we can make this information more accessible? I have reformulated the release text in the hopes that this clears things up a bit already.
