### 2019-02-01 f06e01d
Context diameter factor settings
