### 2019-02-01 a49cd5d
FeatureScout: Add skeleton for setContextDiameterFactor in 3d cylinder vis
