### 2019-02-01 33ba673
FIAKER: Add (empty) volume dataset settings widget
