### 2019-02-01 5c18246
Core: FixedAspectRatio widget vtk 8.2 fixes
