### 2019-02-01 5263f81
FIAKER: More vtk 8.2 fixes
