### 2019-02-03 #1185 @behollister
@trallard, I'm currently using [Linux Build](https://github.com/3dct/open_iA/wiki/Linux-Build) instructions to verify manual installation. Mostly looks okay.

However, I can't verify [Windows Build](https://github.com/3dct/open_iA/wiki/Windows-Build), as I don't have access to a Windows system.

Will it be possible for you or @katyhuff to verify building / installation on Windows?
