### 2019-02-04 #1185 @trallard
I can conduct my review on a Windows machine :grin: if needed but if this works on Linux I would say it satisfies the installation requirements
