### 2019-02-05 3c14d12
FeatureScout: SPM color themes are rendered in 3D view
