+++ core/src/iALUT.cpp
@@ -38,7 +38,10 @@ const QStringList colormaps = QStringList()
 	<< "ColorBrewer single hue 5-class oranges"
 	<< "ColorBrewer single hue 5-class grays"
 	<< "ColorBrewer single hue 5-class oranges inv"
-	<< "ColorBrewer single hue 5-class red inv";
+	<< "ColorBrewer single hue 5-class red inv"
+	<< "ColorBrewer qualitative 12-class Set3"
+	<< "ColorBrewer qualitative 9-class Set1"
+	<< "ColorBrewer qualitative 12-class Paired";
 
 const QStringList& iALUT::GetColorMapNames()
 {
@@ -138,7 +141,7 @@ int iALUT::BuildLUT( vtkSmartPointer<vtkLookupTable> pLUT, double * lutRange, QS
 		c.setRgb(253, 190, 133); ctf->AddRGBPoint(0.25, c.redF(), c.greenF(), c.blueF());
 		c.setRgb(253, 141, 60);	 ctf->AddRGBPoint(0.5, c.redF(), c.greenF(), c.blueF());
 		c.setRgb(230, 85, 13);	 ctf->AddRGBPoint(0.75, c.redF(), c.greenF(), c.blueF());
-		c.setRgb(166, 54, 3);	 ctf->AddRGBPoint(0.0, c.redF(), c.greenF(), c.blueF());
+		c.setRgb(166, 54, 3);	 ctf->AddRGBPoint(1.0, c.redF(), c.greenF(), c.blueF());
 		break;
 
 	case 9:
@@ -147,7 +150,52 @@ int iALUT::BuildLUT( vtkSmartPointer<vtkLookupTable> pLUT, double * lutRange, QS
 		c.setRgb(252, 174, 145); ctf->AddRGBPoint(0.25, c.redF(), c.greenF(), c.blueF());
 		c.setRgb(251, 106, 74);	 ctf->AddRGBPoint(0.5, c.redF(), c.greenF(), c.blueF());
 		c.setRgb(222, 45, 38);	 ctf->AddRGBPoint(0.75, c.redF(), c.greenF(), c.blueF());
-		c.setRgb(165, 15, 21);	 ctf->AddRGBPoint(0.0, c.redF(), c.greenF(), c.blueF());
+		c.setRgb(165, 15, 21);	 ctf->AddRGBPoint(1.0, c.redF(), c.greenF(), c.blueF());
+		break;
+
+	case 10:
+		// ColorBrewer qualitative 12-class Set3
+		c.setRgb(141, 211, 199); ctf->AddRGBPoint(0.0, c.redF(), c.greenF(), c.blueF());
+		c.setRgb(255, 255, 179); ctf->AddRGBPoint(0.09, c.redF(), c.greenF(), c.blueF());
+		c.setRgb(190, 186, 218); ctf->AddRGBPoint(0.18, c.redF(), c.greenF(), c.blueF());
+		c.setRgb(251, 128, 114); ctf->AddRGBPoint(0.27, c.redF(), c.greenF(), c.blueF());
+		c.setRgb(128, 177, 211); ctf->AddRGBPoint(0.36, c.redF(), c.greenF(), c.blueF());
+		c.setRgb(253, 180, 98);  ctf->AddRGBPoint(0.45, c.redF(), c.greenF(), c.blueF());
+		c.setRgb(179, 222, 105); ctf->AddRGBPoint(0.54, c.redF(), c.greenF(), c.blueF());
+		c.setRgb(252, 205, 229); ctf->AddRGBPoint(0.63, c.redF(), c.greenF(), c.blueF());
[…]
