+++ modules/FeatureScout/iAFeatureScoutSPLOM.cpp
@@ -76,6 +76,7 @@ void iAFeatureScoutSPLOM::initScatterPlot(QDockWidget* container, vtkTable* csvT
 	matrix->settings.enableColorSettings = true;
 	connect(matrix, &iAQSplom::selectionModified, this, &iAFeatureScoutSPLOM::selectionModified);
 	connect(matrix, &iAQSplom::parameterVisibilityChanged, this, &iAFeatureScoutSPLOM::parameterVisibilityChanged);
+	connect(matrix, SIGNAL(lookupTableChanged()), this, SLOT(lookupTableChanged()));
 
 	QAction* addClass = new QAction(QObject::tr("Add class"), nullptr);
 	connect(addClass, &QAction::triggered, this, &iAFeatureScoutSPLOM::addClass);
@@ -105,6 +106,11 @@ void iAFeatureScoutSPLOM::setParameterVisibility(size_t paramIndex, bool visible
 {
 	if (matrix)
 		matrix->setParameterVisibility(paramIndex, visible);
+}
+
+void iAFeatureScoutSPLOM::lookupTableChanged()
+{
+	emit renderLUTChanges(matrix->lookupTable(), matrix->colorLookupParam());
 }
 
 void iAFeatureScoutSPLOM::setFilter(int classID)
