### 2019-02-06 #1185 @behollister
Currently having issues with various itk header files during compilation. One example:

[ 40%] Building CXX object modules/CMakeFiles/Segmentation.dir/Segmentation/iAWatershedSegmentation.cpp.o
In file included from /home/brad/JOSS/workspace/src/Toolkit/FuzzyCMeans/itkFCMClassifierInitializationImageFilter.h:21:0,
                 from /home/brad/JOSS/workspace/src/modules/Segmentation/iAFuzzyCMeans.cpp:23:
/home/brad/JOSS/workspace/src/Toolkit/FuzzyCMeans/itkFuzzyClassifierInitializationImageFilter.h:31:10: fatal error: itkMultiThreader.h: No such file or directory
 #include "itkMultiThreader.h"
