### 2019-02-06 #1185 @behollister
ITK_DIR points to /bin-4.13-itk, as directed per [build instructions](https://github.com/3dct/open_iA/wiki/Linux-Build).

"Set ITK_DIR: /workspace/ITK/bin-4.13.0"

Will see about setting path to header files as needed for Linux build...
