### 2019-02-06 8147c47
Uncertainty: Initialize toolbar pointer
