### 2019-02-06 31e13bb
FIAKER: shrink diameter only of context fibers (requires vtkTubeFilter changes)
