+++ modules/FiAKEr/iAFiAKErController.cpp
@@ -46,6 +46,7 @@
 #include "qthelper/iADockWidgetWrapper.h"
 #include "iALookupTable.h"
 #include "iALUT.h"
+#include "iAMapperImpl.h"
 #include "iAMathUtility.h"
 #include "iAModuleDispatcher.h"
 #include "iARendererManager.h"
@@ -104,9 +105,12 @@ namespace
 	int HistogramBins = 20;
 	int SelectionOpacity = iA3DLineObjectVis::DefaultSelectionOpacity;
 	int ContextOpacity = iA3DLineObjectVis::DefaultContextOpacity;
+	const double MinDiameterFactor = 0.01;
+	const double MaxDiameterFactor = 1.5;
+	const int MinFactorSliderVal = 1;
+	const int MaxFactorSliderVal = 100;
 	double DiameterFactor = 1.0;
 	double ContextDiameterFactor = 1.0;
-	const double DiameterFactorDivisor = 10; // = 1, value is divided / 10
 	const size_t NoPlotsIdx = std::numeric_limits<size_t>::max();
 	const size_t NoResult = NoPlotsIdx;
 	const QString ModuleSettingsKey("FIAKER");
@@ -317,19 +321,20 @@ QWidget* iAFiAKErController::setupSettingsView()
 	connect(m_contextOpacitySlider, &QSlider::valueChanged, this, &iAFiAKErController::contextOpacityChanged);
 	m_contextOpacityLabel = new QLabel(QString::number(ContextOpacity));
 
+	m_diameterFactorMapper = new iALogarithmicMapper(MinDiameterFactor, MaxDiameterFactor, 1, 100);
 	auto diameterFactorSlider = new QSlider(Qt::Horizontal);
-	diameterFactorSlider->setMinimum(1);
-	diameterFactorSlider->setMaximum(static_cast<int>(DiameterFactorDivisor*DiameterFactorDivisor));
-	diameterFactorSlider->setValue(static_cast<int>(DiameterFactor * DiameterFactorDivisor));
+	diameterFactorSlider->setMinimum(MinFactorSliderVal);
+	diameterFactorSlider->setMaximum(MaxFactorSliderVal);
+	diameterFactorSlider->setValue(static_cast<int>(m_diameterFactorMapper->srcToDst(DiameterFactor)));
 	connect(diameterFactorSlider, &QSlider::valueChanged, this, &iAFiAKErController::diameterFactorChanged);
-	m_diameterFactorLabel = new QLabel(QString::number(1.0));
+	m_diameterFactorLabel = new QLabel(QString::number(DiameterFactor));
 	
 	auto contextDiameterFactorSlider = new QSlider(Qt::Horizontal);
-	contextDiameterFactorSlider->setMinimum(1);
-	contextDiameterFactorSlider->setMaximum(static_cast<int>(DiameterFactorDivisor * DiameterFactorDivisor));
-	contextDiameterFactorSlider->setValue(static_cast<int>(ContextDiameterFactor * DiameterFactorDivisor));
+	contextDiameterFactorSlider->setMinimum(MinFactorSliderVal);
+	contextDiameterFactorSlider->setMaximum(MaxFactorSliderVal);
+	contextDiameterFactorSlider->setValue(static_cast<int>(m_diameterFactorMapper->srcToDst(ContextDiameterFactor)));
 	connect(contextDiameterFactorSlider, &QSlider::valueChanged, this, &iAFiAKErController::contextDiameterFactorChanged);
-	m_contextDiameterFactorLabel = new QLabel(QString::number(1.0));
+	m_contextDiameterFactorLabel = new QLabel(QString::number(ContextDiameterFactor));
 
[…]
