+++ modules/FeatureScout/iA3DCylinderObjectVis.cpp
@@ -33,7 +33,10 @@
 
 iA3DCylinderObjectVis::iA3DCylinderObjectVis( iAVtkWidget* widget, vtkTable* objectTable, QSharedPointer<QMap<uint, uint> > columnMapping,
 	QColor const & color, int numberOfCylinderSides ):
-	iA3DLineObjectVis( widget, objectTable, columnMapping, color )
+	iA3DLineObjectVis( widget, objectTable, columnMapping, color ),
+	m_objectCount(objectTable->GetNumberOfRows()),
+	m_contextFactors(nullptr),
+	m_contextDiameterFactor(1.0)
 {
 	auto tubeRadius = vtkSmartPointer<vtkDoubleArray>::New();
 	tubeRadius->SetName("TubeRadius");
@@ -77,4 +80,36 @@ void iA3DCylinderObjectVis::setDiameterFactor(double diameterFactor)
 
 void iA3DCylinderObjectVis::setContextDiameterFactor(double contextDiameterFactor)
 {
+	if (contextDiameterFactor == 1.0)
+	{
+		if (m_contextFactors)
+			delete m_contextFactors;
+		m_contextFactors = nullptr;
+	}
+	else
+	{
+		if (!m_contextFactors)
+			m_contextFactors = new float[2 * m_objectCount];
+		m_contextDiameterFactor = contextDiameterFactor;
+		size_t selIdx = 0;
+		for (vtkIdType row = 0; row < m_objectCount; ++row)
+		{
+			bool isSelected = selIdx < m_selection.size() && (m_selection[selIdx] == row);
+			if (isSelected)
+				++selIdx;
+			float diameter = (!isSelected) ? m_contextDiameterFactor : 1.0;
+			m_contextFactors[2 * row] = diameter;
+			m_contextFactors[2 * row + 1] = diameter;
+		}
+	}
+	m_tubeFilter->SetIndividualFactors(m_contextFactors);
+	m_tubeFilter->Modified();
+	m_tubeFilter->Update();
+	updateRenderer();
+}
+
+void iA3DCylinderObjectVis::setSelection(std::vector<size_t> const & sortedSelInds, bool selectionActive)
+{
+	iA3DLineObjectVis::setSelection(sortedSelInds, selectionActive);
+	setContextDiameterFactor(m_contextDiameterFactor);
 }
