### 2019-02-07 24d0932
FIAKER: improve Factor mapping and slider value display
