+++ modules/FiAKEr/iAFiAKErController.cpp
@@ -105,8 +105,8 @@ namespace
 	int HistogramBins = 20;
 	int SelectionOpacity = iA3DLineObjectVis::DefaultSelectionOpacity;
 	int ContextOpacity = iA3DLineObjectVis::DefaultContextOpacity;
-	const double MinDiameterFactor = 0.01;
-	const double MaxDiameterFactor = 1.5;
+	const double MinDiameterFactor = 0.02;
+	const double MaxDiameterFactor = 2;
 	const int MinFactorSliderVal = 1;
 	const int MaxFactorSliderVal = 100;
 	double DiameterFactor = 1.0;
@@ -312,29 +312,31 @@ QWidget* iAFiAKErController::setupSettingsView()
 	m_defaultOpacitySlider->setMaximum(255);
 	m_defaultOpacitySlider->setValue(SelectionOpacity);
 	connect(m_defaultOpacitySlider, &QSlider::valueChanged, this, &iAFiAKErController::mainOpacityChanged);
-	m_defaultOpacityLabel = new QLabel(QString::number(SelectionOpacity));
+	m_defaultOpacityLabel = new QLabel(QString::number(SelectionOpacity, 'f', 2));
 
 	m_contextOpacitySlider = new QSlider(Qt::Horizontal);
 	m_contextOpacitySlider->setMinimum(0);
 	m_contextOpacitySlider->setMaximum(255);
 	m_contextOpacitySlider->setValue(ContextOpacity);
 	connect(m_contextOpacitySlider, &QSlider::valueChanged, this, &iAFiAKErController::contextOpacityChanged);
-	m_contextOpacityLabel = new QLabel(QString::number(ContextOpacity));
+	m_contextOpacityLabel = new QLabel(QString::number(ContextOpacity, 'f', 2));
 
-	m_diameterFactorMapper = new iALogarithmicMapper(MinDiameterFactor, MaxDiameterFactor, 1, 100);
+	m_diameterFactorMapper = new iALinearMapper(MinDiameterFactor, MaxDiameterFactor, MinFactorSliderVal, MaxFactorSliderVal);
 	auto diameterFactorSlider = new QSlider(Qt::Horizontal);
 	diameterFactorSlider->setMinimum(MinFactorSliderVal);
 	diameterFactorSlider->setMaximum(MaxFactorSliderVal);
-	diameterFactorSlider->setValue(static_cast<int>(m_diameterFactorMapper->srcToDst(DiameterFactor)));
+	int factorSliderValue = static_cast<int>(m_diameterFactorMapper->srcToDst(DiameterFactor));
+	diameterFactorSlider->setValue(factorSliderValue);
 	connect(diameterFactorSlider, &QSlider::valueChanged, this, &iAFiAKErController::diameterFactorChanged);
-	m_diameterFactorLabel = new QLabel(QString::number(DiameterFactor));
+	m_diameterFactorLabel = new QLabel(QString::number(m_diameterFactorMapper->dstToSrc(factorSliderValue), 'f', 2));
 	
 	auto contextDiameterFactorSlider = new QSlider(Qt::Horizontal);
 	contextDiameterFactorSlider->setMinimum(MinFactorSliderVal);
 	contextDiameterFactorSlider->setMaximum(MaxFactorSliderVal);
-	contextDiameterFactorSlider->setValue(static_cast<int>(m_diameterFactorMapper->srcToDst(ContextDiameterFactor)));
+	int contextFactorSlider = static_cast<int>(m_diameterFactorMapper->srcToDst(ContextDiameterFactor));
+	contextDiameterFactorSlider->setValue(contextFactorSlider);
[…]
