### 2019-02-07 dc033f3
Create custom iAvtkTubeFilter copied from VTK to preserve diameter factor modifications
