+++ modules/FeatureScout/iAvtkTubeFilter.h
@@ -1,0 +1,241 @@
+/*
+* Adapted version of vtkTubeFilter.h from the Visualization toolkit, taken
+* from v8.2. For full license and original source code, please refer to
+* the VTK source archive from vtk.org.
+*
+* In comparison to the vtkTubeFilter included in vtk 8.x, it contains two
+* functional changes to VTK_VARY_RADIUS_BY_ABSOLUTE_SCALAR mode:
+*    - the RadiusFactor is applied to the radii retrieved from the scalars
+*    - a separate IndividualFactors array can be set, with additional
+*      diameter adaptation factors for each single point
+*/
+/*=========================================================================
+
+  Copyright (c) Ken Martin, Will Schroeder, Bill Lorensen
+  All rights reserved.
+  See Copyright.txt or http://www.kitware.com/Copyright.htm for details.
+
+=========================================================================*/
+#pragma once
+
+#include <vtkFiltersCoreModule.h> // For export macro
+#include <vtkPolyDataAlgorithm.h>
+
+#define VTK_VARY_RADIUS_OFF 0
+#define VTK_VARY_RADIUS_BY_SCALAR 1
+#define VTK_VARY_RADIUS_BY_VECTOR 2
+#define VTK_VARY_RADIUS_BY_ABSOLUTE_SCALAR 3
+
+#define VTK_TCOORDS_OFF                    0
+#define VTK_TCOORDS_FROM_NORMALIZED_LENGTH 1
+#define VTK_TCOORDS_FROM_LENGTH            2
+#define VTK_TCOORDS_FROM_SCALARS           3
+
+class vtkCellArray;
+class vtkCellData;
+class vtkDataArray;
+class vtkFloatArray;
+class vtkPointData;
+class vtkPoints;
+
+class iAvtkTubeFilter : public vtkPolyDataAlgorithm
+{
+public:
+  vtkTypeMacro(iAvtkTubeFilter,vtkPolyDataAlgorithm);
+  void PrintSelf(ostream& os, vtkIndent indent) override;
+
+  void SetIndividualFactors(float* indivFactors);
+
+  /**
+   * Construct object with radius 0.5, radius variation turned off, the
+   * number of sides set to 3, and radius factor of 10.
+   */
+  static iAvtkTubeFilter *New();
+
+  //@{
+  /**
+   * Set the minimum tube radius (minimum because the tube radius may vary).
+   */
+  vtkSetClampMacro(Radius,double,0.0,VTK_DOUBLE_MAX);
+  vtkGetMacro(Radius,double);
+  //@}
+
+  //@{
+  /**
+   * Turn on/off the variation of tube radius with scalar value.
+   */
+  vtkSetClampMacro(VaryRadius,int,
+                   VTK_VARY_RADIUS_OFF,VTK_VARY_RADIUS_BY_ABSOLUTE_SCALAR);
+  vtkGetMacro(VaryRadius,int);
+  void SetVaryRadiusToVaryRadiusOff()
+    {this->SetVaryRadius(VTK_VARY_RADIUS_OFF);};
+  void SetVaryRadiusToVaryRadiusByScalar()
+    {this->SetVaryRadius(VTK_VARY_RADIUS_BY_SCALAR);};
+  void SetVaryRadiusToVaryRadiusByVector()
+    {this->SetVaryRadius(VTK_VARY_RADIUS_BY_VECTOR);};
+  void SetVaryRadiusToVaryRadiusByAbsoluteScalar()
+    {this->SetVaryRadius(VTK_VARY_RADIUS_BY_ABSOLUTE_SCALAR);};
+  const char *GetVaryRadiusAsString();
+  //@}
+
+  //@{
+  /**
[…]
