### 2019-02-07 #1185 @behollister
I'm using 18.04.1 LTS. ITK / VTK compiled fine per the instructions.

As I installed the ITK / VTK builds, CMake picked up those locations. When the build failed, I switched ITK_DIR to the build directory (then re-ran cmake after a 'make clean' in the build dir).

Best that I start anew, following instructions *exactly*. Also, I've found that reconfig sometimes fails. Cleaning out the previous build often fixes issues.

Will work through this again on Friday. Thank you for the suggestions. Sorry for any unnecessary delays.
