### 2019-02-08 #1185 @codeling
What we've been thinking about for some time is providing a "superbuild", i.e. a CMake project that gathers, configures and builds all the required libraries for open_iA. I'm currently trying to put something together in that direction. This should make it easier to set up a working build environment.
Maybe you want to wait on that; though I'm not sure when exactly I will be able to provide this superbuild.
