### 2019-02-08 3d3dc4c
FeatureScout: Use QSharedPointer instead of ref to iALookupTable
