+++ modules/FeatureScout/dlg_FeatureScout.h
@@ -135,7 +135,7 @@ private slots:
 	void saveStl();
 	void updateStlProgress(int i);
 	void updateMarProgress(int i);
-	void renderLUTChanges(iALookupTable const &lut, int colInd);
+	void renderLUTChanges(QSharedPointer<iALookupTable> lut, int colInd);
 private:
 	//create labelled output image based on defined classes
 	template <class T> void CreateLabelledOutputMask(iAConnector & con, const QString & fOutPath);
