### 2019-02-08 f27e4da
Morphology Filters: Fill Hole Filter added, Opening and Closing filter are now based on ITK's grayscale version of these filters (convince and performance).
