+++ modules/Morphology/iAMorphologyFilters.h
@@ -25,5 +25,6 @@
 IAFILTER_DEFAULT_CLASS(iADilation);
 IAFILTER_DEFAULT_CLASS(iAErosion);
 IAFILTER_DEFAULT_CLASS(iAVesselEnhancement);
-IAFILTER_DEFAULT_CLASS(iAOpening);
-IAFILTER_DEFAULT_CLASS(iAClosing);
+IAFILTER_DEFAULT_CLASS(iAMorphOpening);
+IAFILTER_DEFAULT_CLASS(iAMorphClosing);
+IAFILTER_DEFAULT_CLASS(iAFillHole);
