### 2019-02-08 72877a2
CMake: Fix cmake-version dependant C++ Standard setting
