+++ cmake/Modules/Common.cmake
@@ -510,16 +510,15 @@ IF (MSVC)
 	ADD_DEFINITIONS(-D_SCL_SECURE_NO_WARNINGS)
 ENDIF (MSVC)
 IF (CMAKE_COMPILER_IS_GNUCXX)
-	# sets the c++17 standard (or the next lower one which is available on the used cmake/compiler)
-	IF (CMAKE_VERSION VERSION_GREATER "3.10")
-		MESSAGE(STATUS "Aiming for C++20 support.")
-		SET(CMAKE_CXX_STANDARD 20)
-	ELSEIF (CMAKE_VERSION VERSION_GREATER "3.7")
+	IF (CMAKE_VERSION VERSION_LESS "3.8")
+		MESSAGE(STATUS "Aiming for C++14 support.")
+		SET(CMAKE_CXX_STANDARD 14)
+	ELSEIF (CMAKE_VERSION VERSION_LESS "3.11")
 		MESSAGE(STATUS "Aiming for C++17 support.")
 		SET(CMAKE_CXX_STANDARD 17)
 	ELSE()
-		MESSAGE(STATUS "Aiming for C++14 support.")
-		SET(CMAKE_CXX_STANDARD 14)
+		MESSAGE(STATUS "Aiming for C++20 support.")
+		SET(CMAKE_CXX_STANDARD 20)
 	ENDIF()
 	SET(CMAKE_CXX_EXTENSIONS OFF)
 	# Make sure at least C++ 0x is supported:
