### 2019-02-08 #1185 @behollister
I'll give it another try before Monday as well. Btw, the Blender project provides a "superbuild" script. You might use that as a starting point.

See:
https://developer.blender.org/diffusion/B/browse/master/build_files/build_environment/install_deps.sh
