### 2019-02-08 #1185 @behollister
> There is a somewhat working version of a superbuild available now, see https://github.com/3dct/open_iA-superbuild ; so far we have tested it only on Ubuntu 18.04.1 and 18.10, it should simplify the build procedure significantly! Hope this helps.

Yes, this should be helpful. Will try this soon!
