### 2019-02-11 f90e036
FeatureScout: tensor calculation fix
