### 2019-02-21 #1185 @behollister
Superbuild works fine on my system. Trying to verify functionality now...

@codeling, is there an automated test suite & data for each area of functionality claimed in https://github.com/3dct/open_iA/wiki/User-Guide? I don't see this provided as a script in the source tree, or any installation of sample data.

@katyhuff, what degree of coverage is needed for testing functionality? For example, each use case for a given data set could itself be tested with each file type supported:
https://github.com/3dct/open_iA/wiki/File-Formats
