### 2019-02-22 #1185 @codeling
At this point, we only have limited automatic tests available for some library functions.
See the [iASimpleTester](https://github.com/3dct/open_iA/blob/master/core/src/iASimpleTester.h), used in tests of the [iAStringHelper](https://github.com/3dct/open_iA/blob/master/core/src/iAStringHelperTest.cpp) and in two tests in the [RandomWalker module](https://github.com/3dct/open_iA/tree/master/modules/RandomWalker).

We were planning to include more tests but haven't found the time to do so yet.

The existing tests are run as part of our CDash build. [Its latest results are available here](https://git.3dct.at/CDash/index.php?project=openiA&date=2019-02-12). Typically this build runs daily, but at the moment the machine running the build is turned off due to remodeling in our offices, it will be online again on Monday.

Since we recently added the possibility to run the filters in the modules via command line, we are planning to add tests for all filters and file formats there.

Most of our visual analytics tools are heavily GUI- and interaction-based. We therefore currently do not test these automatically, but would be very glad about pointers how to easily achieve that!
