### 2019-02-22 c1bda04
Cmd: Fix filter max output
