### 2019-02-27 5185fc1
iAVec3 -> iAVec3f, add doxygen comments, introduce methods to unify with iAVec3 from 4DCT module
