+++ Toolkit/Dreamcaster/dreamcaster.cpp
@@ -692,7 +692,7 @@ void DreamCaster::RenderViewsSlot()
 	set_pos[1] = ui.sb_posy->value()/stngs.SCALE_COEF;
 	set_pos[2] = ui.sb_posz->value()/stngs.SCALE_COEF;
 	tracer->setPositon(set_pos);
-	iAVec3 transl = -iAVec3(set_pos);
+	iAVec3f transl = -iAVec3f(set_pos);
 	tracer->GetScene()->recalculateD(&transl);
 	int paramIndex = ui.cb_rangeParameter->currentIndex();
 	tracer->SetCutAABBList(&cutFigList->aabbs);
@@ -700,9 +700,9 @@ void DreamCaster::RenderViewsSlot()
 	int totalRends = cntX*cntY*cntZ;
 	if(ui.cb_RadonSA->currentIndex() == 2)//do only Radon space analysis
 	{
-		iAVec3 l_o; // rays' origin point
-		iAVec3 l_vp_corners[2];// plane's corners in 3d
-		iAVec3 l_vp_delta[2];// plane's x and y axes' directions in 3D
+		iAVec3f l_o; // rays' origin point
+		iAVec3f l_vp_corners[2];// plane's corners in 3d
+		iAVec3f l_vp_delta[2];// plane's x and y axes' directions in 3D
 		int counter = 0;
 		for (int x=0; x<cntX; x++)
 		{
@@ -767,18 +767,18 @@ void DreamCaster::RenderViewsSlot()
 		int counter = 0;
 		unsigned int batch_counter=0;
 		//batch parameters for every render
-		iAVec3 * s1_o = new iAVec3[stngs.BATCH_SIZE]; 
-		iAVec3 * corners = new iAVec3[stngs.BATCH_SIZE];
-		iAVec3 * dxs = new iAVec3[stngs.BATCH_SIZE];
-		iAVec3 * dys = new iAVec3[stngs.BATCH_SIZE];
+		iAVec3f * s1_o = new iAVec3f[stngs.BATCH_SIZE]; 
+		iAVec3f * corners = new iAVec3f[stngs.BATCH_SIZE];
+		iAVec3f * dxs = new iAVec3f[stngs.BATCH_SIZE];
+		iAVec3f * dys = new iAVec3f[stngs.BATCH_SIZE];
 		unsigned int * xs = new unsigned int[stngs.BATCH_SIZE];//prev skipped positions when batching
 		unsigned int * ys = new unsigned int[stngs.BATCH_SIZE];//prev skipped positions when batching
 		unsigned int * zs = new unsigned int[stngs.BATCH_SIZE];//prev skipped positions when batching
 		float * rotsX = new float[stngs.BATCH_SIZE];
 		float * rotsY = new float[stngs.BATCH_SIZE]; 
 		float * rotsZ = new float[stngs.BATCH_SIZE]; 
-		iAVec3 s1_vp_corners[2];
-		iAVec3 s1_vp_delta[2];
+		iAVec3f s1_vp_corners[2];
+		iAVec3f s1_vp_delta[2];
 		unsigned int s1_x = 0;
 		unsigned int s1_y = 0;
 		unsigned int s1_z = 0;
@@ -949,9 +949,9 @@ void DreamCaster::RenderViewsSlot()
 	}//GPU
 	else//using CPU
 	{
-		iAVec3 o; // rays' origin point
-		iAVec3 vp_corners[2];// plane's corners in 3d
-		iAVec3 vp_delta[2];// plane's x and y axes' directions in 3D
+		iAVec3f o; // rays' origin point
+		iAVec3f vp_corners[2];// plane's corners in 3d
+		iAVec3f vp_delta[2];// plane's x and y axes' directions in 3D
[…]
