### 2019-02-27 #1185 @katyhuff
Sorry for the delay in responding to your question, @behollister .

> what degree of coverage is needed for testing functionality? For example, each use case for a given data set could itself be tested with each file type supported: https://github.com/3dct/open_iA/wiki/File-Formats

Generally speaking, my approach to this is to assess whether the scientific functionality of the work is demonstrated and confimed by the tests, so full coverage is certainly not required, but test coverage should confirm the capability claims of the work.

Anyway -- thanks for conducting your review @behollister !
