### 2019-02-27 a99eea6
JOSS: bib capitalization
