+++ JOSS-paper/paper.bib
@@ -47,7 +47,7 @@
 }
 
 @article{Reh:2015,
-  title     = {Fuzzy feature tracking: Visual analysis of industrial 4D-XCT data},
+  title     = {Fuzzy feature tracking: Visual analysis of industrial {4D-XCT} data},
   author    = {Andreas Reh and Artem Amirkhanov and Christoph Heinzl and Johann Kastner and Eduard Gr{\"o}ller},
   journal   = {Computers {\&} Graphics},
   volume    = {3},
@@ -83,7 +83,7 @@
 }
 
 @article{Reh:2013,
-  title     = {MObjects -- {A} Novel Method for the Visualization and Interactive Exploration of Defects in Industrial XCT Data},
+  title     = {MObjects -- {A} Novel Method for the Visualization and Interactive Exploration of Defects in Industrial {XCT} Data},
   author    = {Andreas Reh and Christian Gusenbauer and Johann Kastner and Eduard Gr{\"o}ller and Christoph Heinzl},
   journal   = {IEEE Transactions on Visualization and Computer Graphics},
   volume    = {19},
