### 2019-02-28 f90fdec
Core: Fix Qt < 5.10 build (shortCutVisbileInContextMenu)
