### 2019-03-01 86ea7ca
Core: Additional Qt < 5.10 build fix
