### 2019-03-05 80d6780
Update RenderSettings - slice plane opacity
