### 2019-03-05 e61cdfe
renaming slice plane opacity
