### 2019-03-05 ca5f992
WireFrame for extraction of volume
