+++ core/src/iALUT.cpp
@@ -198,9 +198,15 @@ int iALUT::BuildLUT( vtkSmartPointer<vtkLookupTable> pLUT, double const * lutRan
 		c.setRgb(177, 89, 40); ctf->AddRGBPoint(1.0, c.redF(), c.greenF(), c.blueF());
 		break;
 	}
-
+#if VTK_MAJOR_VERSION < 8 || (VTK_MAJOR_VERSION == 8 && VTK_MINOR_VERSION == 0)
+	double lutRangeNonConst[2];
+	std::copy(lutRange, lutRange + 2, lutRangeNonConst);
+	pLUT->SetRange(lutRangeNonConst);
+	pLUT->SetTableRange(lutRangeNonConst);
+#else
 	pLUT->SetRange( lutRange );
 	pLUT->SetTableRange( lutRange );
+#endif
 	pLUT->SetNumberOfColors( numCols );
 	for( int i = 0; i < numCols; ++i )
 	{
