### 2019-03-05 ceb477e
Support origins other than 0,0,0
