### 2019-03-05 6b92ebd
Core: Minor rename refactoring
