### 2019-03-06 b7c3d81
Accept files to open via drag&drop
