+++ core/src/mainwindow.cpp
@@ -43,12 +43,16 @@
 #include <vtkImageData.h>
 #include <vtkOpenGLRenderer.h>
 #include <vtkPiecewiseFunction.h>
+#include <vtkSmartVolumeMapper.h>
 #include <vtkVersion.h>
 
 #include <QCloseEvent>
+#include <QDragEnterEvent>
+#include <QDropEvent>
 #include <QComboBox>
 #include <QFileDialog>
 #include <QMessageBox>
+#include <QMimeData>
 #include <QMdiSubWindow>
 #include <QSettings>
 #include <QSignalMapper>
@@ -65,6 +69,7 @@ MainWindow::MainWindow(QString const & appName, QString const & version, QString
 	m_gitVersion(version)
 {
 	setupUi(this);
+	setAcceptDrops(true);
 
 	// restore geometry and state
 	QCoreApplication::setOrganizationName("FHW");
@@ -129,7 +134,6 @@ MainWindow::MainWindow(QString const & appName, QString const & version, QString
 	statusBar()->showMessage(tr("Ready"));
 }
 
-
 MainWindow::~MainWindow()
 {
 	// save geometry and state
@@ -142,13 +146,11 @@ MainWindow::~MainWindow()
 	windowMapper = 0;
 }
 
-
 void MainWindow::timeout()
 {
 	splashScreen->finish(this);
 	delete timer;
 }
-
 
 bool MainWindow::KeepOpen()
 {
@@ -173,7 +175,6 @@ bool MainWindow::KeepOpen()
 	return false;
 }
 
-
 void MainWindow::closeEvent(QCloseEvent *event)
 {
 	if (KeepOpen())
@@ -191,6 +192,22 @@ void MainWindow::closeEvent(QCloseEvent *event)
 	}
 }
 
+void MainWindow::dragEnterEvent(QDragEnterEvent *e)
+{
+	if (e->mimeData()->hasUrls())
+	{
+		e->acceptProposedAction();
+	}
+}
+
+void MainWindow::dropEvent(QDropEvent *e)
+{
+	for(const QUrl &url: e->mimeData()->urls())
+	{
+		QString fileName = url.toLocalFile();
+		LoadFile(fileName);
+	}
+}
 
 void MainWindow::CloseAllSubWindows()
 {
@@ -199,7 +216,6 @@ void MainWindow::CloseAllSubWindows()
 		mdiArea->closeAllSubWindows();
 	}
 }
-
 
 void MainWindow::Open()
 {
@@ -212,7 +228,6 @@ void MainWindow::Open()
 		)
 	);
 }
-
 
 void MainWindow::OpenRaw()
 {
@@ -240,7 +255,6 @@ void MainWindow::OpenRaw()
 	}
 }
 
-
 void MainWindow::OpenImageStack()
 {
 	LoadFile(
@@ -252,7 +266,6 @@ void MainWindow::OpenImageStack()
 		), true
 	);
 }
-
 
 void MainWindow::OpenVolumeStack()
 {
@@ -266,7 +279,6 @@ void MainWindow::OpenVolumeStack()
 	);
 }
 
-
 void MainWindow::OpenRecentFile()
 {
 	QAction *action = qobject_cast<QAction *>(sender());
@@ -276,7 +288,6 @@ void MainWindow::OpenRecentFile()
 	QString fileName = action->data().toString();
 	LoadFile(fileName);
 }
-
 
 void MainWindow::LoadFile(QString const & fileName)
 {
@@ -290,7 +301,6 @@ void MainWindow::LoadFile(QString const & fileName)
 		LoadFile(fileName, fileName.endsWith(".volstack"));
 	}
 }
-
 
 void MainWindow::LoadFile(QString fileName, bool isStack)
 {
[…]
