+++ modules/GEMSe/dlg_samplingSettings.cpp
@@ -31,7 +31,6 @@
 #include <iANameMapper.h>
 
 #include <QCheckBox>
-#include <QDropEvent>
 #include <QFileDialog>
 #include <QMimeData>
 #include <QShortcut>
