### 2019-03-06 586effa
AstraReconstruction: Option to initialize SIRT/CGLS with FDK
