### 2019-03-06 6de3ca0
AstraReconstruction: Only execute initial FDK if SIRT/CGLS is configured
