### 2019-03-07 3dct/open_iA#32 @katyhuff
@trallard : have you checked out the binaries mentioned above? The linux binary works for me, as does the superbuild. (but I'm not on windows...)
