### 2019-03-07 3dct/open_iA#36 @katyhuff
[JOSS REVIEW] Minor build error in paper.md
Whedon is having trouble with the current paper.md.

```

Dload Upload Total Spent Left Speed

Error producing PDF.
! Missing $ inserted.

$
l.358 T

Looks like we failed to compile the PDF
```

Please see comment here for what I believe to be the solution:
https://github.com/3dct/open_iA/commit/e02c5481f2d2a14719a2c25c17f7733477d5f134#r32637253
