### 2019-03-07 #1185 @katyhuff
https://github.com/3dct/open_iA/issues/36
