### 2019-03-07 #1185 @katyhuff
@codeling I'm going through this a bit to bring this submission to completion. Please see: https://github.com/3dct/open_iA/issues/36
