### 2019-03-07 6a9d40c
JOSS: Test if removing the link fixes the build
