### 2019-03-07 3dct/open_iA#36 @codeling
Unfortunately that was not it. Not even removing the link completely fixes the issue

I'm pretty sure it's actually not a change on our side that causes the build to fail, as the paper.md that was used for the first build today was not changed since the last generated pdf (https://github.com/openjournals/joss-reviews/issues/1185#issuecomment-458211753). My local test build with pandoc also succeeds. I have no idea what the issue could be here. Has maybe something changed recently in the way that whedon builds the pdf? Or is there any way to get more comprehensive error output for the build?
