### 2019-03-08 2b61649
renamed transparency -> opacity slice planes
