### 2019-03-08 6815c26
FeatureScout: Move commonly applicable code from LineObjectVis to ColoredPolyObjectVis
