### 2019-03-08 e198feb
FIAKER: Enable Ellipse visualization
