+++ modules/FiAKEr/iAFiAKErController.cpp
@@ -27,12 +27,11 @@
 #include "iAStackedBarChart.h"
 
 // FeatureScout:
+#include "iA3DCylinderObjectVis.h"
+#include "iA3DEllipseObjectVis.h"
 #include "iACsvConfig.h"
 #include "iAFeatureScoutModuleInterface.h"
 #include "iAVectorPlotData.h"
-#include "qthelper/iAFixedAspectWidget.h"
-#include "qthelper/iASignallingWidget.h"
-#include "io/iAFileUtils.h"
 
 // Core:
 #include "charts/iAChartWidget.h"
@@ -44,7 +43,6 @@
 #include "iAColorTheme.h"
 #include "iAConnector.h"
 #include "iAConsole.h"
-#include "qthelper/iADockWidgetWrapper.h"
 #include "iALookupTable.h"
 #include "iALUT.h"
 #include "iAMapperImpl.h"
@@ -52,13 +50,17 @@
 #include "iAModuleDispatcher.h"
 #include "iARendererManager.h"
 #include "iAStringHelper.h"
-#include "io/iAITKIO.h"
 #include "iATransferFunction.h"
 #include "iAVolumeRenderer.h"
 #include "iAVtkWidget.h"
+#include "io/iAITKIO.h"
 #include "io/iAFileChooserWidget.h"
+#include "io/iAFileUtils.h"
 #include "mainwindow.h"
 #include "mdichild.h"
+#include "qthelper/iADockWidgetWrapper.h"
+#include "qthelper/iAFixedAspectWidget.h"
+#include "qthelper/iASignallingWidget.h"
 
 #include <vtkColorTransferFunction.h>
 #include <vtkCubeSource.h>
@@ -153,8 +155,8 @@ class iAFiberCharUIData
 {
 public:
 	iAVtkWidget* vtkWidget;
-	QSharedPointer<iA3DCylinderObjectVis> mini3DVis;
-	QSharedPointer<iA3DCylinderObjectVis> main3DVis;
+	QSharedPointer<iA3DColoredPolyObjectVis> mini3DVis;
+	QSharedPointer<iA3DColoredPolyObjectVis> main3DVis;
 	QCheckBox* cbBoundingBox;
 	QCheckBox* cbShow;
 	iAChartWidget* histoChart;
@@ -640,6 +642,20 @@ QWidget* iAFiAKErController::setupOptimStepView()
 	return optimStepsView;
 }
 
+namespace
+{
+	QSharedPointer<iA3DColoredPolyObjectVis> create3DVis(iAVtkWidget* vtkWidget,
+		vtkSmartPointer<vtkTable> table, QSharedPointer<QMap<uint, uint> > mapping, QColor const & color, int objectType)
+	{
+		switch (objectType)
+		{
+		case iACsvConfig::Ellipses:  return QSharedPointer<iA3DColoredPolyObjectVis>(new iA3DEllipseObjectVis(vtkWidget, table, mapping, color));
+		default:
+		case iACsvConfig::Cylinders: return QSharedPointer<iA3DColoredPolyObjectVis>(new iA3DCylinderObjectVis(vtkWidget, table, mapping, color));
+		}
+	}
+}
+
 QWidget* iAFiAKErController::setupResultListView()
 {
 	size_t commonPrefixLength = 0, commonSuffixLength = 0;
@@ -773,10 +789,8 @@ QWidget* iAFiAKErController::setupResultListView()
 		m_resultsListLayout->addWidget(uiData.stackedBars, resultID + 1, StackedBarColumn);
 		m_resultsListLayout->addWidget(uiData.histoChart, resultID + 1, HistogramColumn);
 
[…]
