### 2019-03-08 #1185 @trallard
Hi sorry for the massive delay in this.
I tested in a Windows machine and all seems to be working ok.  Due to the comments above regarding tests/and the functionality claims of the package I ran my own review and local test based on the user guide.

So far I could not spot anything that was not working as expected or that limited the use of the software for research purposes.

As this develops and to address the concern raised by @behollister

>>@katyhuff, what degree of coverage is needed for testing functionality? For example, each use case for a given data set could itself be tested with each file type supported:
https://github.com/3dct/open_iA/wiki/File-Formats

More examples covering more file formats could be added, though they are not required for acceptance of this piece of software in JOSS.
