### 2019-03-11 8a7f8db
FIAKER: Fix build old Qt (< 5.11)
