### 2019-03-12 b38965a
CPack: Fix install for vtk < 8.2
