+++ cmake/Modules/Common.cmake
@@ -372,8 +372,10 @@ IF (${VTK_MAJOR_VERSION} GREATER 8 OR (${VTK_MAJOR_VERSION} EQUAL 8 AND ${VTK_MI
 ELSE()
 	SET (VTK_EXTRA_LIBS ${VTK_EXTRA_LIBS} vtkalglib vtkexoIIc vtkoggtheora)
 	IF (WIN32)
+		SET (VTK_LIB_DIR "${VTK_DIR}/bin/Release")
 		INSTALL(FILES ${VTK_LIB_DIR}/QVTKWidgetPlugin.dll DESTINATION .)
 	ELSE()
+		SET (VTK_LIB_DIR "${VTK_DIR}/lib")
 		INSTALL(FILES ${VTK_LIB_DIR}/libQVTKWidgetPlugin.so DESTINATION .)
 	ENDIF()
 ENDIF()
