### 2019-03-13 #1185 @codeling
Took me longer than planned to prepare the release, but here it is:

[open_iA 2019.03](https://github.com/3dct/open_iA/releases/tag/2019.03)

DOI: [10.5281/zenodo.2591999](https://doi.org/10.5281/zenodo.2591999)
