### 2019-03-13 #1185 @katyhuff
So, I'm sorry I didn't get back to you earlier on your previous question.

It needs to be the version that was reviewed (so, your master branch...) particularly if changes have been significant. However, a new submission can be made to review the changes since then, and that review should go fasterand will result in another version of the joss DOI.
