### 2019-03-14 7a13765
BoneThickness: Fix interactor Console message on start
