### 2019-03-14 ed0440a
FIAKER: Fix slot connection
