### 2019-03-15 e3a5b3b
Doxygen: Fix warnings, unify commenting style
