+++ Toolkit/Dreamcaster/raycast/include/scene.h
@@ -48,6 +48,8 @@ struct intersection
 * Ray-AABB intersection routine.
 * @param ray ray class.
 * @param box axis aligned bounding box structure.
+* @param tmin parent AABB min t.
+* @param tmax parent AABB max t.
 * @return 
 1 - if ray intersects AABB
 0 - otherwise
@@ -157,6 +159,8 @@ inline int Intersect(iAVec3f & ro, iAVec3f rd, const aabb& box)
 * Ray-Axis-oriented-cylinder intersection routine.
 * @param ray ray class.
 * @param box axis aligned bounding box structure.
+* @param tmin parent AABB min t.
+* @param tmax parent AABB max t.
 * @param ind cylinder axis index.
 * @return 
 1 - if ray intersects AABB
@@ -168,7 +172,7 @@ int IntersectCyl(const Ray & ray, const aabb& box, float &tmin, float&tmax, int 
 * checks which subnodes' AABBs are intersected by ray
 * @param ray ray class.
 * @param tmin parent AABB min t.
-* @param tmin parent AABB max t.
+* @param tmax parent AABB max t.
 * @param split split plane's coordinates
 * @param splitIndex index of splitting axis
 * @param t [out] split plane's t
