### 2019-03-15 906d332
DreamCaster: Replace custom deg2rad/rad2deg by respective vtkMath functions
