+++ Toolkit/Dreamcaster/dreamcaster.cpp
@@ -809,14 +809,14 @@ void DreamCaster::RenderViewsSlot()
 				placementsParams[s1_x][s1_z] = parameters_t();
 			}
 			actor->SetOrientation(0,0,0);
-			actor->RotateWXYZ( rad2deg(minValX) + rad2deg(deltaX)*s1_x	,1,0,0 );//degrees
-			actor->RotateWXYZ( rad2deg(deltaY)*s1_y						,0,1,0 );
-			actor->RotateZ(    rad2deg(minValZ) + rad2deg(deltaZ)*s1_z);
+			actor->RotateWXYZ( vtkMath::DegreesFromRadians(minValX) + vtkMath::DegreesFromRadians(deltaX)*s1_x	,1,0,0 );//degrees
+			actor->RotateWXYZ( vtkMath::DegreesFromRadians(deltaY)*s1_y						,0,1,0 );
+			actor->RotateZ(    vtkMath::DegreesFromRadians(minValZ) + vtkMath::DegreesFromRadians(deltaZ)*s1_z);
 			actor->SetPosition(set_pos[0], set_pos[1], set_pos[2]);
 			cutAABActor->SetOrientation(0,0,0);
-			cutAABActor->RotateWXYZ( rad2deg(minValX) + rad2deg(deltaX)*s1_x	,1,0,0 );//degrees
-			cutAABActor->RotateWXYZ( rad2deg(deltaY)*s1_y						,0,1,0 );
-			cutAABActor->RotateZ(    rad2deg(minValZ) + rad2deg(deltaZ)*s1_z);
+			cutAABActor->RotateWXYZ( vtkMath::DegreesFromRadians(minValX) + vtkMath::DegreesFromRadians(deltaX)*s1_x	,1,0,0 );//degrees
+			cutAABActor->RotateWXYZ( vtkMath::DegreesFromRadians(deltaY)*s1_y						,0,1,0 );
+			cutAABActor->RotateZ(    vtkMath::DegreesFromRadians(minValZ) + vtkMath::DegreesFromRadians(deltaZ)*s1_z);
 			cutAABActor->SetPosition(set_pos[0], set_pos[1], set_pos[2]);
 			double bottom = actor->GetBounds()[2];
 			plateActor->SetPosition(0.0, bottom-0.5*PLATE_HEIGHT, 0.0);
@@ -963,14 +963,14 @@ void DreamCaster::RenderViewsSlot()
 					if(isStopped)
 						return;
 					actor->SetOrientation(0,0,0);
-					actor->RotateWXYZ( rad2deg(minValX) + rad2deg(deltaX)*x	,1,0,0 );//degrees
-					actor->RotateWXYZ(                    rad2deg(deltaY)*y	,0,1,0 );
-					actor->RotateZ   ( rad2deg(minValZ) + rad2deg(deltaZ)*z);
+					actor->RotateWXYZ( vtkMath::DegreesFromRadians(minValX) + vtkMath::DegreesFromRadians(deltaX)*x	,1,0,0 );//degrees
+					actor->RotateWXYZ(                    vtkMath::DegreesFromRadians(deltaY)*y	,0,1,0 );
+					actor->RotateZ   ( vtkMath::DegreesFromRadians(minValZ) + vtkMath::DegreesFromRadians(deltaZ)*z);
 					actor->SetPosition(set_pos[0], set_pos[1], set_pos[2]);
 					cutAABActor->SetOrientation(0,0,0);
[…]
