### 2019-03-15 e9947b3
Core: Add doxygen comments to iAModuleInterface
