### 2019-03-15 fff92e5
Doxygen: improve brief descriptions
