### 2019-03-19 59ba66f
Core Refactor Mainwindow: Consistent connect (+order by menu/toolbar), method name case
