+++ core/src/mainwindow.cpp
@@ -130,7 +130,7 @@ MainWindow::MainWindow(QString const & appName, QString const & version, QString
 	this->layoutToolbar->insertWidget(this->actionSave_Layout, layout);
 
 	m_moduleDispatcher->InitializeModules(iAConsoleLogger::Get());
-	SetModuleActionsEnabled( false );
+	setModuleActionsEnabled( false );
 	statusBar()->showMessage(tr("Ready"));
 }
 
@@ -152,10 +152,10 @@ void MainWindow::timeout()
 	delete timer;
 }
 
-bool MainWindow::KeepOpen()
+bool MainWindow::keepOpen()
 {
 	bool childHasChanges = false;
-	for (MdiChild* mdiChild: MdiChildList())
+	for (MdiChild* mdiChild: mdiChildList())
 		childHasChanges |= mdiChild->isWindowModified();
 	if (childHasChanges)
 	{
@@ -168,7 +168,7 @@ bool MainWindow::KeepOpen()
 		}
 		else
 		{ // avoid individual questions for each window
-			for (MdiChild* mdiChild: MdiChildList())
+			for (MdiChild* mdiChild: mdiChildList())
 				mdiChild->setWindowModified(false);
 		}
 	}
@@ -177,7 +177,7 @@ bool MainWindow::KeepOpen()
 
 void MainWindow::closeEvent(QCloseEvent *event)
 {
-	if (KeepOpen())
+	if (keepOpen())
 	{
 		event->ignore();
 		return;
@@ -205,21 +205,21 @@ void MainWindow::dropEvent(QDropEvent *e)
 	for(const QUrl &url: e->mimeData()->urls())
 	{
 		QString fileName = url.toLocalFile();
-		LoadFile(fileName);
+		loadFile(fileName);
 	}
 }
 
-void MainWindow::CloseAllSubWindows()
+void MainWindow::closeAllSubWindows()
 {
-	if (!KeepOpen())
+	if (!keepOpen())
 	{
 		mdiArea->closeAllSubWindows();
 	}
 }
 
-void MainWindow::Open()
+void MainWindow::open()
 {
-	LoadFiles(
+	loadFiles(
 		QFileDialog::getOpenFileNames(
 			this,
 			tr("Open Files"),
@@ -229,7 +229,7 @@ void MainWindow::Open()
 	);
 }
 
-void MainWindow::OpenRaw()
+void MainWindow::openRaw()
 {
 	QString fileName = QFileDialog::getOpenFileName(
 		this,
@@ -255,9 +255,9 @@ void MainWindow::OpenRaw()
 	}
 }
 
-void MainWindow::OpenImageStack()
+void MainWindow::openImageStack()
 {
-	LoadFile(
+	loadFile(
 		QFileDialog::getOpenFileName(
 			this,
 			tr("Open File"),
@@ -267,9 +267,9 @@ void MainWindow::OpenImageStack()
 	);
 }
 
-void MainWindow::OpenVolumeStack()
+void MainWindow::openVolumeStack()
 {
-	LoadFile(
+	loadFile(
 		QFileDialog::getOpenFileName(
 			this,
 			tr("Open File"),
@@ -279,30 +279,30 @@ void MainWindow::OpenVolumeStack()
 	);
 }
 
-void MainWindow::OpenRecentFile()
+void MainWindow::openRecentFile()
 {
 	QAction *action = qobject_cast<QAction *>(sender());
 	if (!action)
 		return;
 
 	QString fileName = action->data().toString();
-	LoadFile(fileName);
+	loadFile(fileName);
 }
 
-void MainWindow::LoadFile(QString const & fileName)
[…]
