### 2019-03-19 #1185 @katyhuff
Hi @codeling sorry for the confusion. I think we can move forward with just a little clarification, if you have a moment. This will help me identify a path forward quickly, as I am on vacation and won't be able to look closely at the full diff for a few days. I'll try to answer some questions, and will ask a couple as well.

> I'm curious then about the proposed workflow and tagging scheme in a JOSS review: If the version specified at submission and the one at the end of the review (in between which potential changes can occur, especially those triggered by the review) should be the same, this implies that the tag needs to be modified, right?

The DOI created for the JOSS submission should include all commits in the version you originally submitted and the commits comprising changes triggered by the review, which is why we ask for you to create that release at the end of the review, on the branch that we reviewed. Ideally, there shouldn't be lots of other commits folded in that we never saw during review. Some minor changes are generally fine, but I'm confused by your comments about the DOI you made, so I'm not sure what the case is here.

It seems that the intermediate step, required due to your dated releasing system, included folding in other edits that were never involved in the review.  The two comments on this seem to be slightly conflicting about the magnitude of those changes:

> a lot has changed since that version already, so changing that version feels wrong
> Changes between 2018.12 and 2019.03 include those done in response to the review comments, but also a few other adaptations (mostly bugfixes).

But, in looking quickly at the release notes, those changes do indeed seem to be mostly bugfixes and targetted improvements:

- Adaptable slice plane opacity
- Support for volumes with origins other than 0,0,0
- Fixed handling of special characters in file names
- Extract Image: Indication of extracted cube in 3D view
- ASTRA: possibility to initialize SIRT/CGLS with FDK reconstruction

Regarding your comment:
> I'm curious then about the proposed workflow and tagging scheme in a JOSS review... While technically this is no problem, this for me goes a bit against the original intention of tags (permanently naming a specific state).

Usually, submitting authors request a review on master make all triggered changes in that branch, while keeping other major development in a separate branch during review. But, there are a handful of ways to keep major edits out of the doi that is archived alongside the JOSS review (a joss-review branch that is rebased at some point... etc.). The release tag associated with the JOSS review should permanently point to the release that JOSS reviewers reviewed (including changes triggered by the review,  because those were part of the review)

> So I guess we should resubmit the 2019.03 version for a new review in JOSS?
If I and the reviewers can quickly agree that the minor edits and bugfixes do not need their own independent review, then no. But, if those changes are major, then, yes.

Final comment: I'm also somewhat confused about this statement -- can you clarify what you mean here :
> Edit: Updated DOI - using DOI referencing all (future) versions of open_iA
