### 2019-03-19 #1185 @codeling
> Some minor changes are generally fine, but I'm confused by your comments about the DOI you made, so I'm not sure what the case is here.
>It seems that the intermediate step, required due to your dated releasing system, included folding in other edits that were never involved in the review. The two comments on this seem to be slightly conflicting about the magnitude of those changes:

The main issue probably is that I didn't think of keeping master at only what we submitted for JOSS plus the changes resulting from review comments. It would of course have been possible to keep those changes in another branch for the time being. Instead, some bug fixes and minor enhancements were intermixed with the JOSS review changes, sorry for this. Sorry also for the confusing comment about "a lot has changed". To clarify, the changes to the core functionality are actually minor, you probably wouldn't notice them if you hadn't used the program extensively before. However, since we do these dated version tags, it just didn't "feel" right then, to update a tag from December 2018 to point to the version created in March 2019 with those fixes.

Regarding the DOI confusion:

> Edit: Updated DOI - using DOI referencing all (future) versions of open_iA

As far as I got it there are two zenodo DOI's you get:
- One for the specific submitted artifact (the source zip from github), of the software as submitted - this is 10.5281/zenodo.2591999 in this case, and
- Another DOI (10.5281/zenodo.2591998 in this case), which "represents all versions, and will always resolve to the latest one. " according to the "Versions" box on the lower right at https://zenodo.org/record/2591999.

I did the edit when I realized that there was such a non-version-specific DOI, and used it here.
But I suppose for the submission you would need the one for the specific submitted artifact?
