### 2019-03-19 cf0150e
Core: add toolbar-only actions to menu, Unify mainwindow action names
