+++ core/src/mainwindow.cpp
@@ -96,8 +96,8 @@ MainWindow::MainWindow(QString const & appName, QString const & version, QString
 
 	splashScreen->showMessage("\n      Setup UI...", Qt::AlignTop, QColor(255, 255, 255));
 	applyQSS();
-	actionLink_views->setChecked(defaultSlicerSettings.LinkViews);//removed from readSettings, if is needed at all?
-	actionLink_mdis->setChecked(defaultSlicerSettings.LinkMDIs);
+	actionLinkViews->setChecked(defaultSlicerSettings.LinkViews);//removed from readSettings, if is needed at all?
+	actionLinkMdis->setChecked(defaultSlicerSettings.LinkMDIs);
 	setCentralWidget(mdiArea);
 
 	windowMapper = new QSignalMapper(this);
@@ -107,11 +107,11 @@ MainWindow::MainWindow(QString const & appName, QString const & version, QString
 	updateMenus();
 	slicerToolsGroup = new QActionGroup(this);
 	slicerToolsGroup->setExclusive(false);
-	slicerToolsGroup->addAction(actionSnake_Slicer);
+	slicerToolsGroup->addAction(actionSnakeSlicer);
 	slicerToolsGroup->addAction(actionRawProfile);
 
-	actionDelete_point->setEnabled(false);
-	actionChange_color->setEnabled(false);
+	actionDeletePoint->setEnabled(false);
+	actionChangeColor->setEnabled(false);
 
 	splashScreen->showMessage(tr("\n      Version: %1").arg (m_gitVersion), Qt::AlignTop, QColor(255, 255, 255));
 
@@ -127,7 +127,7 @@ MainWindow::MainWindow(QString const & appName, QString const & version, QString
 	this->layout->setStyleSheet("padding: 0");
 	this->layout->resize(this->layout->geometry().width(), 100);
 	this->layout->setSizeAdjustPolicy(QComboBox::AdjustToContents);
-	this->layoutToolbar->insertWidget(this->actionSave_Layout, layout);
+	this->layoutToolbar->insertWidget(this->actionSaveLayout, layout);
 
 	m_moduleDispatcher->InitializeModules(iAConsoleLogger::Get());
 	setModuleActionsEnabled( false );
@@ -563,6 +563,23 @@ bool MainWindow::loadSettings()
 	return true;
 }
 
+namespace
+{
+	void removeNode(QDomNode &rootNode, char const *str)
+	{
+		QDomNodeList list = rootNode.childNodes();
+		for (int n = 0; n < int(list.length()); n++)
+		{
+			QDomNode node = list.item(n);
+			if (node.nodeName() == str)
+			{
+				rootNode.removeChild(node);
+				break;
+			}
+		}
+	}
+}
+
 void MainWindow::saveCamera(QDomDocument &doc)
 {
 	vtkCamera *camera = activeMdiChild()->getRenderer()->GetRenderer()->GetActiveCamera();
@@ -976,20 +993,6 @@ void MainWindow::loadSlicerSettings(QDomNode &slicerSettingsNode)
 	activeMdiChild()->editSlicerSettings(defaultSlicerSettings);
 }
 
-void MainWindow::removeNode(QDomNode &rootNode, char const *str)
-{
-	QDomNodeList list = rootNode.childNodes();
-	for (int n = 0; n < int(list.length()); n++)
-	{
-		QDomNode node = list.item(n);
-		if (node.nodeName() == str)
-		{
-			rootNode.removeChild(node);
-			break;
-		}
-	}
-}
-
 void MainWindow::maxXY()
 {
 	if (activeMdiChild() && activeMdiChild()->xyview())
[…]
