### 2019-03-19 00fd1d6
Core: Minor mainwindow/mdichild method refactoring
