### 2019-03-19 1dbaa96
Core: Fix maximize slicer/renderer if currently other view maximized
