### 2019-03-19 dbfbaea
FIAKER: Add shortcuts for starting
