+++ core/src/io/iAIOProvider.h
@@ -45,4 +45,5 @@ public:
 	static const QString ProjectFileTypeFilter;
 	static const QString ProjectFileExtension;
 	static const QString MetaImages;
+	static const QString VTKFiles; 
 };
