### 2019-03-20 cf81298
Core: Fix crash in Histogram on deleting modality
