### 2019-03-20 a2b0701
Core: Rewrite channels to use created ID instead of fixed IDs from iAChannelID
