### 2019-03-21 5fc0ead
Connected Components: remove 'simple' from class names and description
