### 2019-03-21 0e75f1d
Connected Components: Rename category to 'Connected Components' (remove 'Filters')
