### 2019-03-21 5ee6698
Enable InterActorStyle For Modality
