### 2019-03-21 #1185 @codeling
> sorry for the delay in responding - am still on vacation

Ah yes forgot that - sorry for bothering you on vacation!

> I think this can be accepted with the DOI including your interleaved changes, as they seem mostly minor. I'll move forward with that.

Great, thanks!
