### 2019-03-21 #1185 @katyhuff
Thanks @codeling for your submission patience and cooperation with the process!

Thank you @trallard and @behollister for your detailed reviews!

This paper should be ready to accept @openjournals/joss-eics - over to you!
